### 2022-01-31 #4109 @diehlpk
Hi @arfon I am happy to edit the paper. However, I have to declare that one of the authors and I have contributed code to the same repository a few years ago.
