### 2022-01-31 #4115 @whedon
[REVIEW]: Cabana: A Performance Portable Library for Particle-Based Simulations
**Submitting author:** @streeve (<a href="http://orcid.org/0000-0002-4250-9476">Samuel Reeve</a>)
**Repository:** https://github.com/ECP-copa/Cabana
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.5.0
