+++ doxygen/search/all_c.js
@@ -1,22 +1,22 @@
 var searchData=
 [
-  ['l2g_410',['L2G',['../structCajita_1_1IndexConversion_1_1L2G.html',1,'Cajita::IndexConversion::L2G&lt; MeshType, EntityType &gt;'],['../structCajita_1_1IndexConversion_1_1L2G.html#a02ccb00f4dfcd66076ce37d727e98a53',1,'Cajita::IndexConversion::L2G::L2G()']]],
-  ['label_411',['label',['../classCajita_1_1Array.html#a89f55b2eb65c85bee16a84eca8b6f993',1,'Cajita::Array::label()'],['../classCabana_1_1Slice.html#af22a113cd1d1795e45fe2288d16ed85d',1,'Cabana::Slice::label()'],['../classCabana_1_1AoSoA.html#ae00f8e72faacf696043da28005c57ce5',1,'Cabana::AoSoA::label()']]],
-  ['layout_412',['layout',['../structCajita_1_1LayoutAdapter.html#a0f4c7011975932331bd011ecc73cecf9',1,'Cajita::LayoutAdapter::layout()'],['../classCajita_1_1Array.html#aa5fe3e40a31d3ca1a507d5ef8c90d6de',1,'Cajita::Array::layout()'],['../structCajita_1_1ReferenceConjugateGradient_1_1LayoutContainer.html#a7d306d90d8e94e60fde35bc7de928c03',1,'Cajita::ReferenceConjugateGradient::LayoutContainer::layout()']]],
-  ['layoutadapter_413',['LayoutAdapter',['../structCajita_1_1LayoutAdapter.html',1,'Cajita']]],
-  ['layoutcabanaslice_414',['LayoutCabanaSlice',['../structKokkos_1_1LayoutCabanaSlice.html',1,'Kokkos::LayoutCabanaSlice&lt; SOASTRIDE, VLEN, DIM0, DIM1, DIM2, DIM3, DIM4, DIM5 &gt;'],['../structKokkos_1_1LayoutCabanaSlice.html#a53dc5465c4fd54fdc8182a221cba7b17',1,'Kokkos::LayoutCabanaSlice::LayoutCabanaSlice(LayoutCabanaSlice &amp;&amp;)=default'],['../structKokkos_1_1LayoutCabanaSlice.html#aadfb9de405102f0e3ed7908d78aa503a',1,'Kokkos::LayoutCabanaSlice::LayoutCabanaSlice(LayoutCabanaSlice const &amp;)=default'],['../structKokkos_1_1LayoutCabanaSlice.html#a7b39173a84769cb34e13949f0690a413',1,'Kokkos::LayoutCabanaSlice::LayoutCabanaSlice(size_t num_soa=0, size_t vector_length=VectorLength, size_t d0=D0, size_t d1=D1, size_t d2=D2, size_t d3=D3, size_t d4=D4, size_t d5=D5)']]],
-  ['layoutcontainer_415',['LayoutContainer',['../structCajita_1_1ReferenceConjugateGradient_1_1LayoutContainer.html',1,'Cajita::ReferenceConjugateGradient']]],
[…]
