+++ doxygen/search/classes_f.js
@@ -1,40 +1,40 @@
 var searchData=
 [
-  ['scalargradientg2p_895',['ScalarGradientG2P',['../structCajita_1_1ScalarGradientG2P.html',1,'Cajita']]],
-  ['scalargradientp2g_896',['ScalarGradientP2G',['../structCajita_1_1ScalarGradientP2G.html',1,'Cajita']]],
-  ['scalarvalueg2p_897',['ScalarValueG2P',['../structCajita_1_1ScalarValueG2P.html',1,'Cajita']]],
-  ['scalarvaluep2g_898',['ScalarValueP2G',['../structCajita_1_1ScalarValueP2G.html',1,'Cajita']]],
-  ['secondneighborstag_899',['SecondNeighborsTag',['../classCabana_1_1SecondNeighborsTag.html',1,'Cabana']]],
-  ['serialoptag_900',['SerialOpTag',['../classCabana_1_1SerialOpTag.html',1,'Cabana']]],
-  ['simdpolicy_901',['SimdPolicy',['../classCabana_1_1SimdPolicy.html',1,'Cabana']]],
-  ['slice_902',['Slice',['../classCabana_1_1Slice.html',1,'Cabana']]],
-  ['soa_903',['SoA',['../structCabana_1_1SoA.html',1,'Cabana']]],
-  ['soa_3c_20membertypes_3c_20types_2e_2e_2e_20_3e_2c_201_20_3e_904',['SoA&lt; MemberTypes&lt; Types... &gt;, 1 &gt;',['../structCabana_1_1SoA.html',1,'Cabana']]],
-  ['soa_3c_20membertypes_3c_20types_2e_2e_2e_20_3e_2c_20vectorlength_20_3e_905',['SoA&lt; MemberTypes&lt; Types... &gt;, VectorLength &gt;',['../structCabana_1_1SoA_3_01MemberTypes_3_01Types_8_8_8_01_4_00_01VectorLength_01_4.html',1,'Cabana']]],
-  ['sparsedimpartitioner_906',['SparseDimPartitioner',['../classCajita_1_1SparseDimPartitioner.html',1,'Cajita']]],
-  ['sparsemap_907',['SparseMap',['../classCajita_1_1SparseMap.html',1,'Cajita']]],
-  ['sparsemesh_908',['SparseMesh',['../structCajita_1_1SparseMesh.html',1,'Cajita']]],
-  ['spline_909',['Spline',['../structCajita_1_1Spline.html',1,'Cajita']]],
-  ['spline_3c_200_20_3e_910',['Spline&lt; 0 &gt;',['../structCajita_1_1Spline_3_010_01_4.html',1,'Cajita']]],
-  ['spline_3c_201_20_3e_911',['Spline&lt; 1 &gt;',['../structCajita_1_1Spline_3_011_01_4.html',1,'Cajita']]],
-  ['spline_3c_202_20_3e_912',['Spline&lt; 2 &gt;',['../structCajita_1_1Spline_3_012_01_4.html',1,'Cajita']]],
-  ['spline_3c_203_20_3e_913',['Spline&lt; 3 &gt;',['../structCajita_1_1Spline_3_013_01_4.html',1,'Cajita']]],
-  ['splinedata_914',['SplineData',['../structCajita_1_1SplineData.html',1,'Cajita']]],
[…]
