+++ doxygen/search/variables_11.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['x_1412',['x',['../structCajita_1_1SplineDataMember_3_01Scalar_00_01Order_00_01NumSpaceDim_00_01SplineLogicalPosition_01_4.html#a3785c2ef793411791661336b1abb2186',1,'Cajita::SplineDataMember&lt; Scalar, Order, NumSpaceDim, SplineLogicalPosition &gt;']]]
+  ['x_1428',['x',['../structCajita_1_1SplineDataMember_3_01Scalar_00_01Order_00_01NumSpaceDim_00_01SplineLogicalPosition_01_4.html#a3785c2ef793411791661336b1abb2186',1,'Cajita::SplineDataMember&lt; Scalar, Order, NumSpaceDim, SplineLogicalPosition &gt;']]]
 ];
