+++ doxygen/search/classes_11.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['uniformdimpartitioner_943',['UniformDimPartitioner',['../classCajita_1_1UniformDimPartitioner.html',1,'Cajita']]],
-  ['uniformmesh_944',['UniformMesh',['../structCajita_1_1UniformMesh.html',1,'Cajita']]]
+  ['uniformdimpartitioner_951',['UniformDimPartitioner',['../classCajita_1_1UniformDimPartitioner.html',1,'Cajita']]],
+  ['uniformmesh_952',['UniformMesh',['../structCajita_1_1UniformMesh.html',1,'Cajita']]]
 ];

[…]
