### 2022-02-04 ECP-copa/Cabana#496 @XzzX
documentation/examples for interfaces to third-party libraries
It is hard to figure out how to use the interfaces for `heFFTe`, `HYPRE`, `ALL` and `ArborX`. I think the best sources are currently the unit tests. Can you add short usage examples and/or a dedicated wiki page for these interfaces?
