### 2022-02-04 ECP-copa/Cabana#497 @XzzX
add non CPU backend to the CI
Currently all CI jobs target CPU backends of Kokkos. Since GitHub runners do not have GPUs a full GPU CI is not possible this way. However, just compiling with, for example, `nvcc` catches something like `calling host function from device function`.

One of the premade docker containers from Nvidia could do the job.
