### 2022-02-04 #4115 @XzzX
# Reviewer report

## Comments
The software and paper fulfill almost all criteria from the checklist. The software is freely available on GitHub, uses a state-of-the-art build system (CMake), includes examples and unit tests, is tested regularly in multiple configurations via GitHub Actions, is documented via wiki pages available in the GitHub repository and contains a readme, open-source licence and contributions guidelines. I only have minor remarks regarding the CI, documentation and state of the field.

### CI
Currently the CI only tests the CPU backends of Cabana. Since one of the key features of Cabana is its architecture independence at least one other architecture should be tested. Unfortunately, GitHub only provides CPU runners, so GPU tests would be restricted to just compiling. https://github.com/ECP-copa/Cabana/issues/497

### Documentation
The completeness of the documentation varies between the different parts of the software. Whereas the documentation for Cabana is solid, the documentation for the Cajita subpart is incomplete https://github.com/ECP-copa/Cabana/issues/352 and the third-party interfaces documentation is lacking https://github.com/ECP-copa/Cabana/issues/496. However, unit tests are provided for all these parts and some information can be extracted from them.

### State of the field
The paper mentions other software frameworks that can be interfaced with Cabana. However, there is no discussion of competing frameworks or frameworks that cannot be interfaced with Cabana. Since Cabana extends Kokkos with often needed functionality for particle simulations, the framework belongs to the Kokkos ecosystem. Within this ecosystem I am not aware of any competing framework. For the reader all other open-source particle frameworks are potential alternatives to Cabana. In my opinion, a detailed discussion of the alternatives is out of scope for this paper. Perhaps naming a few frameworks with high scalability and support for different backends would be an option. @diehlpk

## Conclusion
Cabana is a valuable addition to the Kokkos ecosystem. It allows faster prototyping and delivers robust performance for fully fledged applications and frameworks. I would really like to see more usage examples / documentation for the advanced functionality of Cabana. Apart from that, the software is in very good shape and the paper should definitely be published.
