### 2022-02-04 f29da97
Add explicit execution type options; update comments
