+++ paper/paper.bib
@@ -191,3 +191,22 @@
     Title = {{ALL - A loadbalancing library, C++ / Fortran library}},
     Url = {http://slms.pages.jsc.fz-juelich.de/websites/all-website},
 }
+
+@article{autopas,
+  title={N ways to simulate short-range particle systems: Automated algorithm selection with the node-level library AutoPas},
+  author={Gratl, Fabio Alexander and Seckler, Steffen and Bungartz, Hans-Joachim and Neumann, Philipp},
+  journal={Computer Physics Communications},
+  pages={108262},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@article{openfpm,
+  title={OpenFPM: A scalable open framework for particle and particle-mesh codes on parallel computers},
+  author={Incardona, Pietro and Leo, Antonio and Zaluzhnyi, Yaroslav and Ramaswamy, Rajesh and Sbalzarini, Ivo F},
+  journal={Computer Physics Communications},
+  volume={241},
+  pages={155--177},
+  year={2019},
+  publisher={Elsevier}
+}
