### 2022-02-07 #4115 @diehlpk
> ### State of the field
>
> The paper mentions other software frameworks that can be interfaced with Cabana. However, there is no discussion of competing frameworks or frameworks that cannot be interfaced with Cabana. Since Cabana extends Kokkos with often needed functionality for particle simulations, the framework belongs to the Kokkos ecosystem. Within this ecosystem I am not aware of any competing framework. For the reader all other open-source particle frameworks are potential alternatives to Cabana. In my opinion, a detailed discussion of the alternatives is out of scope for this paper. Perhaps naming a few frameworks with high scalability and support for different backends would be an option. @diehlpk

I would say mention other frameworks with GPU support would be a good addition. Maybe discuss the supported hardware and explain why this framework is different.
