### 2022-02-09 #4115 @streeve
> I would say mention other frameworks with GPU support would be a good addition. Maybe discuss the supported hardware and explain why this framework is different.

We've added citations to a couple nice general frameworks and are happy to add further discussion as needed
