### 2022-02-10 b0635b6
Remove non-Kokkos versions
