### 2022-02-10 ECP-copa/Cabana#496 @streeve
For each dependency, let's add an in-repo example and add to/update the wiki page (linked from https://github.com/ECP-copa/Cabana/wiki/Programming-Guide):

- @sfogerty heFFTe
- @sslattery hypre
- @aetx ALL
- I'll do ArborX (very little to be done anyway)
