+++ cajita/src/Cajita_Halo.hpp
@@ -748,7 +748,7 @@ class Halo
     {
         auto pp = Cabana::makeParameterPack( array_views... );
         Kokkos::parallel_for(
-            "pack_buffer",
+            "Cajita::Halo::pack_buffer",
             Kokkos::RangePolicy<ExecutionSpace>( exec_space, 0,
                                                  steering.extent( 0 ) ),
             KOKKOS_LAMBDA( const int i ) {
@@ -890,7 +890,7 @@ class Halo
     {
         auto pp = Cabana::makeParameterPack( array_views... );
         Kokkos::parallel_for(
-            "unpack_buffer",
+            "Cajita::Halo::unpack_buffer",
             Kokkos::RangePolicy<ExecutionSpace>( exec_space, 0,
                                                  steering.extent( 0 ) ),
             KOKKOS_LAMBDA( const int i ) {
