+++ example/core_tutorial/CMakeLists.txt
@@ -19,7 +19,7 @@ if(Kokkos_ENABLE_SERIAL)
   add_subdirectory(05_slice)
   add_subdirectory(07_sorting)
   add_subdirectory(08_linked_cell_list)
-  add_subdirectory(09_verlet_list)
+  add_subdirectory(09_neighbor_list)
   if(Kokkos_ENABLE_CUDA)
     add_subdirectory(06_deep_copy)
   endif()
