+++ example/core_tutorial/CMakeLists.txt
@@ -20,6 +20,9 @@ if(Kokkos_ENABLE_SERIAL)
   add_subdirectory(07_sorting)
   add_subdirectory(08_linked_cell_list)
   add_subdirectory(09_neighbor_list)
+  if(Cabana_ENABLE_ARBORX)
+    add_subdirectory(09_neighbor_list_arborx)
+  endif()
   if(Kokkos_ENABLE_CUDA)
     add_subdirectory(06_deep_copy)
   endif()
