### 2022-02-22 ECP-copa/Cabana#502 @atmyers
Broken links to tutorials in wiki
There are some broken links to tutorial files on the wiki, for example on this page: https://github.com/ECP-copa/Cabana/wiki/Sorting-and-Binning

It looks like the tutorials were split into "core" and "cajita" categories, but these links were not updated.
