### 2022-02-22 #4115 @atmyers
Overall, this is a well-written paper about a well-designed and well-tested open source library and I certainly think it should be published in JOSS. I have a few comments that the authors may wish to address:

## Documentation

Both the doxygen API documentation and the wiki are helpful, though I second @XzzX's remarks that the particle-mesh subcomponents in Cajita seem under-covered relative to the core functionality. The linked proxy applications provide a useful intermediary between the relatively simple tests and tutorials and full-fledged production code. I also noticed a few instances of link rot in the wiki section. ECP-copa/Cabana/issues/502

## Benchmarks

The wiki also includes a benchmarks section. The results there are useful and interesting, however, I'm not sure that they fully confirm the performance claims made about the software, because they are (I think) single core and CPU-only and thus don't speak either to performance portability or to scalability. It would be good to include some information about GPU performance and/or scaling results, either in the paper or in the benchmarks section of the wiki, or to cite another source where this information has been included.
