+++ example/cajita_tutorial/05_fast_fourier_transform/fast_fourier_transform_example.cpp
@@ -1,0 +1,203 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+#include <Kokkos_Core.hpp>
+#include <iostream>
+#include <random>
+
+//---------------------------------------------------------------------------//
+// Fast Fourier Transform example.
+//---------------------------------------------------------------------------//
+void FastFourierTransformExample()
+{
+    /*
+      This example shows how to perform a fast fourier transform with Cajita.
+      The current Cajita FFTs take advantage of the heFFTe library, though
+      in the future we plan to support more FFT libraries.
+
+      The basic steps for performing FFTs in Cajita are:
+
+      1. Create a mesh
+
+      2. Create the complex data vector you would like to transform.
+
+      3. Create the FFT and set any options you would like (detailed more below)
+*/
+
+    int comm_rank = -1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &comm_rank );
+
+    if ( comm_rank == 0 )
+    {
+        std::cout << "Cajita Fast Fourier Transform Example\n" << std::endl;
+        std::cout << "    (intended to be run with MPI)\n" << std::endl;
+    }
+
+    /*
+       Create the global mesh.
+       In this example we create a uniform 3D periodic mesh, which is not
+       symmetric. Details on creating a global mesh, grid, and partitioner can
+       be found in previous Cajita examples
+
+       Declare the device memory and execution space to use
+       In this example, we use the host space for execution with OpenMP
+    */
+    using MemorySpace = Kokkos::HostSpace;
+    using ExecutionSpace = Kokkos::DefaultHostExecutionSpace;
+    using DeviceType = Kokkos::Device<ExecutionSpace, MemorySpace>;
+
+    /*
+       Declare which FFT backend to use.
+       Cajita (through heFFTe) has the following backend FFT libraries
+       available:
+       FFTW (default for host execution)
+       cufft (default with Cuda execution)
+       mkl   (default with MKL execution)
+       rocfft (default with HIP execution)
+       In this example we use the default FFT backend type based on the
+       execution space and enabled heFFTe backends, but you could explicitly
+       set an appropriate backend such as Cajita::Experimental::FFTBackendFFTW
+    */
+    using FFTBackendType = Cajita::Experimental::Impl::FFTBackendDefault;
+
+    double cell_size = 0.1;
+    std::array<bool, 3> is_dim_periodic = { true, true, true };
+    std::array<double, 3> global_low_corner = { -1.0, -2.0, -1.0 };
+    std::array<double, 3> global_high_corner = { 1.0, 1.0, 0.5 };
+    auto global_mesh = Cajita::createUniformGlobalMesh(
+        global_low_corner, global_high_corner, cell_size );
+
+    // Create the global grid.
+    Cajita::DimBlockPartitioner<3> partitioner;
+    auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                                 is_dim_periodic, partitioner );
+
+    // Create a local grid.
+    auto local_grid = Cajita::createLocalGrid( global_grid, 0 );
[…]
