### 2022-03-02 3ca762f
Add cajita example folders
