+++ example/cajita_tutorial/05_index_space/CMakeLists.txt
@@ -1,0 +1,14 @@
+############################################################################
+# Copyright (c) 2018-2021 by the Cabana authors                            #
+# All rights reserved.                                                     #
+#                                                                          #
+# This file is part of the Cabana library. Cabana is distributed under a   #
+# BSD 3-clause license. For the licensing terms see the LICENSE file in    #
+# the top-level directory.                                                 #
+#                                                                          #
+# SPDX-License-Identifier: BSD-3-Clause                                    #
+############################################################################
+
+add_executable(IndexSpace index_space_example.cpp)
+target_link_libraries(IndexSpace Cajita)
+add_test(NAME Cajita_tutorial_05 COMMAND ${NONMPI_PRECOMMAND} IndexSpace)
