+++ example/cajita_tutorial/09_grid_parallel/grid_parallel_example.cpp
@@ -1,0 +1,207 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include <iostream>
+
+//---------------------------------------------------------------------------//
+// Grid parallel example.
+//---------------------------------------------------------------------------//
+void gridParallelExample()
+{
+    /*
+      Just as the core library extends the Kokkos::parallel_for and reduce
+      concepts for the AoSoA and neighbor lists, Cajita provides capabilities
+      for parallel iteration over structured grids.
+
+      Most use cases will likely use the interface which use the local grids
+      directly, but there are additional options for using index spaces,
+      including multiple index spaces.
+
+      All data structures (global mesh/grid, local grid/mesh, and array data) is
+      created first.
+    */
+    std::cout << "Cajita Grid Parallel Example\n" << std::endl;
+
+    using exec_space = Kokkos::DefaultHostExecutionSpace;
+    using device_type = exec_space::device_type;
+
+    // Let MPI compute the partitioning for this example.
+    Cajita::UniformDimPartitioner partitioner;
+
+    // Create the global mesh.
+    double cell_size = 0.23;
+    std::array<int, 3> global_num_cell = { 39, 42, 55 };
+    std::array<bool, 3> is_dim_periodic = { false, true, true };
+    std::array<double, 3> global_low_corner = { 1.2, 3.3, -2.8 };
+    std::array<double, 3> global_high_corner = {
+        global_low_corner[0] + cell_size * global_num_cell[0],
+        global_low_corner[1] + cell_size * global_num_cell[1],
+        global_low_corner[2] + cell_size * global_num_cell[2] };
+    auto global_mesh = Cajita::createUniformGlobalMesh(
+        global_low_corner, global_high_corner, global_num_cell );
+
+    // Create the global grid.
+    auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                                 is_dim_periodic, partitioner );
+
+    // Create an array layout on the cells.
+    int halo_width = 2;
+    int dofs_per_cell = 4;
+    auto cell_layout = Cajita::createArrayLayout(
+        global_grid, halo_width, dofs_per_cell, Cajita::Cell() );
+    auto local_grid = cell_layout->localGrid();
+
+    // Create an array.
+    std::string label( "example_array" );
+    auto array = Cajita::createArray<double, device_type>( label, cell_layout );
+    //-----------------------------------------------------------------------//
+
+    /*
+      The grid parallel functions first take a kernel label, then an execution
+      space rather than a direct Kokkos range policy. Internally, a
+      multidimensional range policy (Kokkos::MDRangePolicy) is created with the
+      execution space given. Next, the Own or Ghost tag is passed to define
+      whether the kernel should include ghosted entities or not (remember that
+      specifying Ghost includes both owned and ghosted entities). The mesh
+      entity is passed next and finally, the parallel kernel.
+
[…]
