+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -1,0 +1,231 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+#include <Kokkos_Core.hpp>
+
+#include <iostream>
+
+//---------------------------------------------------------------------------//
+// Interpolation example.
+//---------------------------------------------------------------------------//
+void interpolationExample()
+{
+    /*
+      Cajita provides various particle-to-grid (p2g) and grid-to-particle (g2p)
+      interpolation methods, based on the rank of the interpolated field.
+    */
+    std::cout << "Cajita Interpolation Example\n" << std::endl;
+
+    /*******************************************************************
+      First, we need some setup to demonstrate the use of Cajita interpolation.
+      This includes the creation of a simple uniform mesh and various
+      fields on particles and the mesh.
+    */
+
+    using ExecutionSpace = Kokkos::DefaultHostExecutionSpace;
+
+    // Create the global mesh.
+    std::array<double, 2> low_corner = { -1.2, 0.1 };
+    std::array<double, 2> high_corner = { -0.2, 9.5 };
+    double cell_size = 0.1;
+    auto global_mesh =
+        Cajita::createUniformGlobalMesh( low_corner, high_corner, cell_size );
+
+    // Create the global grid.
+    Cajita::DimBlockPartitioner<2> partitioner;
+    std::array<bool, 2> is_dim_periodic = { true, true };
+    auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                                 is_dim_periodic, partitioner );
+
+    // Create a  grid local_grid.
+    int halo_width = 1;
+    auto local_grid = Cajita::createLocalGrid( global_grid, halo_width );
+    auto local_mesh = Cajita::createLocalMesh<ExecutionSpace>( *local_grid );
+
+    // Create a point in the center of every cell.
+    auto cell_space = local_grid->indexSpace( Cajita::Own(), Cajita::Cell(),
+                                              Cajita::Local() );
+    int num_point = cell_space.size();
+    Kokkos::View<double* [2], ExecutionSpace> points(
+        Kokkos::ViewAllocateWithoutInitializing( "points" ), num_point );
+    Kokkos::parallel_for(
+        "fill_points", createExecutionPolicy( cell_space, ExecutionSpace() ),
+        KOKKOS_LAMBDA( const int i, const int j ) {
+            int pi = i - halo_width;
+            int pj = j - halo_width;
+            int pid = pi + cell_space.extent( Cajita::Dim::I ) * pj;
+            int idx[2] = { i, j };
+            double x[2];
+            local_mesh.coordinates( Cajita::Cell(), idx, x );
+            points( pid, Cajita::Dim::I ) = x[Cajita::Dim::I];
+            points( pid, Cajita::Dim::J ) = x[Cajita::Dim::J];
+        } );
+
+    // Here, we use Cajita functionality to create grid data fields.
+    // Create a scalar field on the grid. See the Array tutorial example for
[…]
