### 2022-03-07 #4115 @streeve
Hey, I fell a little behind on this. ECP-copa/Cabana/issues/502 has been addressed and (for both reviewers) improving documentation is under way in ECP-CoPA/Cabana/issues/#352 and ECP-CoPA/Cabana/issues/#496

You're right that the benchmarks section of the wiki is also out of date with the repo; in addition to the CPU-only AoSoA performance we will add results from the main benchmark kernels on Summit. ECP-CoPA/Cabana/issues/#489 tracks update to the repo, but this will mainly be an update to the wiki
