+++ example/cajita_tutorial/05_index_space/index_space_example.cpp
@@ -1,0 +1,167 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+
+#include <iostream>
+
+//---------------------------------------------------------------------------//
+// Index Space example.
+//---------------------------------------------------------------------------//
+void indexSpaceExample()
+{
+    /*
+      Each index space holds the information on how to iterate over the grid
+      in parallel. It is generally used to thread over a local grid (discussed
+      in the following example) and used tightly coupled to the parallel
+      operations (also discussed in a later example).
+    */
+    std::cout << "Cajita Index Space Example\n" << std::endl;
+
+    /*
+      Index spaces can be one dimensional and up to 4D, mapping to physical
+      dimensions on the grid or grid data. They can be constructed with
+      initializer_lists or arrays.
+
+      In this very simple 1D case, the indices should got from 0-5
+    */
+    Cajita::IndexSpace<1> is1( { 5 } );
+    std::cout << "1D index space:\nMin: ";
+    std::cout << is1.min( 0 ) << " Max: " << is1.max( 0 ) << "\n" << std::endl;
+
+    /*
+      Next, we set both the start and end values for the index space resulting
+      in 5-10.
+    */
+    std::cout << "1D index space:\nMin: ";
+    Cajita::IndexSpace<1> is1_2( { 5 }, { 10 } );
+    std::cout << is1_2.min( 0 ) << " Max: " << is1_2.max( 0 ) << "\n"
+              << std::endl;
+
+    /*
+      Next we create a 2D space.
+    */
+    Cajita::IndexSpace<2> is2( { 4, 3 }, { 8, 9 } );
+    std::cout << "2D index space:\nMin: ";
+    for ( int d = 0; d < 2; ++d )
+        std::cout << is2.min( d ) << " ";
+    std::cout << std::endl << "Max: ";
+    for ( int d = 0; d < 2; ++d )
+        std::cout << is2.max( d ) << " ";
+    std::cout << "\n" << std::endl;
+
+    /*
+      For this 3D case, each dimension should again go from 0 to the value
+      given.
+    */
+    Cajita::IndexSpace<3> is3( { 5, 4, 8 } );
+    std::cout << "3D index space:\nMin: ";
+    for ( int d = 0; d < 3; ++d )
+        std::cout << is3.min( d ) << " ";
+    std::cout << std::endl << "Max: ";
+    for ( int d = 0; d < 3; ++d )
+        std::cout << is3.max( d ) << " ";
+    std::cout << "\n" << std::endl;
+
+    /*
+      There are additional interfaces to extract information
+      about a given index space: rank (dimension), extent and
+      range (pair of start and end) in each dimension, total
+      size of the index space (product).
+    */
+    std::cout << "Rank: " << is3.rank() << std::endl;
+    std::cout << "Extent: ";
+    for ( int d = 0; d < 3; ++d )
+        std::cout << is3.extent( d ) << " ";
+    std::cout << std::endl << "Range: ";
+    for ( int d = 0; d < 3; ++d )
+        std::cout << "(" << is3.range( d ).first << ", "
+                  << is3.range( d ).second << ") ";
+    std::cout << std::endl;
[…]
