+++ example/cajita_tutorial/CMakeLists.txt
@@ -13,3 +13,4 @@ add_subdirectory(01_types)
 add_subdirectory(02_global_mesh)
 add_subdirectory(03_partitioner)
 add_subdirectory(04_global_grid)
+add_subdirectory(05_index_space)
