### 2022-03-08 02c540f
Enable using sharedIndexSpace with local_mesh coord/measure
