+++ cajita/unit_test/tstLocalMesh3d.hpp
@@ -461,6 +461,55 @@ void uniformLocalMeshTest3d( const LocalMeshType& local_mesh,
                     EXPECT_FLOAT_EQ( measure_m( i, j, k ), cell_size );
                 }
     }
+
+    // Extra check for using local mesh functions with shared index space.
+    auto shared_cell_space =
+        local_grid.sharedIndexSpace( Cajita::Own(), Cajita::Cell(), 0, 1, 0 );
+    {
+        auto measure = createView<double, TEST_DEVICE>( "measure", cell_space );
+        auto loc_x = createView<double, TEST_DEVICE>( "loc_x", cell_space );
+        auto loc_y = createView<double, TEST_DEVICE>( "loc_y", cell_space );
+        auto loc_z = createView<double, TEST_DEVICE>( "loc_z", cell_space );
+        Kokkos::parallel_for(
+            "get_cell_coord",
+            createExecutionPolicy( shared_cell_space, TEST_EXECSPACE() ),
+            KOKKOS_LAMBDA( const int i, const int j, const int k ) {
+                Kokkos::Array<double, 3> loc;
+                auto idx = shared_cell_space.min();
+                // Passing this index to the local mesh is the relevant check.
+                local_mesh.coordinates( Cell(), idx.data(), loc.data() );
+                loc_x( i, j, k ) = loc[Dim::I];
+                loc_y( i, j, k ) = loc[Dim::J];
+                loc_z( i, j, k ) = loc[Dim::K];
+                measure( i, j, k ) = local_mesh.measure( Cell(), idx.data() );
+            } );
+        auto measure_m =
+            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace(), measure );
+        auto loc_x_m =
+            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace(), loc_x );
+        auto loc_y_m =
+            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace(), loc_y );
+        auto loc_z_m =
+            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace(), loc_z );
+        for ( int i = 0; i < shared_cell_space.extent( Dim::I ); ++i )
+            for ( int j = 0; j < shared_cell_space.extent( Dim::J ); ++j )
+                for ( int k = 0; k < shared_cell_space.extent( Dim::K ); ++k )
+                {
+                    int i_shared = shared_cell_space.min( Dim::I ) + i;
+                    int j_shared = shared_cell_space.min( Dim::J ) + j;
+                    int k_shared = shared_cell_space.min( Dim::K ) + k;
+
+                    // This is specific to the neighbor at (0,1,0).
+                    EXPECT_FLOAT_EQ( loc_x_m( i_shared, j_shared, k_shared ),
+                                     own_lc_m( Dim::I ) + cell_size * 0.5 );
+                    EXPECT_FLOAT_EQ( loc_y_m( i_shared, j_shared, k_shared ),
+                                     own_hc_m( Dim::J ) - cell_size * 0.5 );
+                    EXPECT_FLOAT_EQ( loc_z_m( i_shared, j_shared, k_shared ),
+                                     own_lc_m( Dim::K ) + cell_size * 0.5 );
+                    EXPECT_FLOAT_EQ( measure_m( i_shared, j_shared, k_shared ),
+                                     cell_size * cell_size * cell_size );
+                }
+    }
 }
 
 //---------------------------------------------------------------------------//
