### 2022-03-09 bff20dc
Add local grid example
