+++ example/cajita_tutorial/06_local_grid/local_grid_example.cpp
@@ -1,0 +1,214 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include <iostream>
+
+#include <mpi.h>
+
+//---------------------------------------------------------------------------//
+// Local Grid example.
+//---------------------------------------------------------------------------//
+void localGridExample()
+{
+    /*
+      The local grid is a subset of the global grid, determined by the
+      partioning of the total system. Just as the global grid defines
+      indexing for the entire domain, the local grid defines indexing for the
+      local MPI rank only, as well as indexing of ghosted portions of the grid
+      for MPI communication. The local grid is the "main" data class in the
+      Cajita subpackage as application users will likely interact with it the
+      most and it includes interfaces to all other grid/mesh classes.
+    */
+    int comm_rank = -1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &comm_rank );
+
+    if ( comm_rank == 0 )
+    {
+        std::cout << "Cajita Local Grid Example" << std::endl;
+        std::cout << "    (intended to be run with MPI)\n" << std::endl;
+    }
+
+    // Here we partition only in x to simplify the example below.
+    int comm_size;
+    MPI_Comm_size( MPI_COMM_WORLD, &comm_size );
+    std::array<int, 3> input_ranks_per_dim = { comm_size, 1, 1 };
+    Cajita::ManualBlockPartitioner<3> partitioner( input_ranks_per_dim );
+
+    // Create the global mesh.
+    double cell_size = 0.23;
+    std::array<int, 3> global_num_cell = { 47, 38, 53 };
+    std::array<double, 3> global_low_corner = { 1.2, 3.3, -2.8 };
+    std::array<double, 3> global_high_corner = {
+        global_low_corner[0] + cell_size * global_num_cell[0],
+        global_low_corner[1] + cell_size * global_num_cell[1],
+        global_low_corner[2] + cell_size * global_num_cell[2] };
+    auto global_mesh = Cajita::createUniformGlobalMesh(
+        global_low_corner, global_high_corner, global_num_cell );
+
+    // Create the global grid.
+    std::array<bool, 3> is_dim_periodic = { true, true, true };
+    auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                                 is_dim_periodic, partitioner );
+
+    /*
+      We create the local grid from the global grid and a halo width -
+      the number of cells communicated with MPI neighbor ranks. The halo width
+      can be queried later if needed.
+    */
+    int halo_width = 2;
+    auto local_grid = Cajita::createLocalGrid( global_grid, halo_width );
+    std::cout << "Minimum halo cell width: " << local_grid->haloCellWidth()
+              << std::endl;
+
+    /*
+      The local grid also stores which MPI ranks are neighbors in the grid (26
+      in 3D, 8 in 2D), using indexing relative to the current local domain. A
[…]
