+++ example/cajita_tutorial/CMakeLists.txt
@@ -14,3 +14,4 @@ add_subdirectory(02_global_mesh)
 add_subdirectory(03_partitioner)
 add_subdirectory(04_global_grid)
 add_subdirectory(05_index_space)
+add_subdirectory(06_local_grid)
