+++ example/cajita_tutorial/CMakeLists.txt
@@ -15,3 +15,4 @@ add_subdirectory(03_partitioner)
 add_subdirectory(04_global_grid)
 add_subdirectory(05_index_space)
 add_subdirectory(06_local_grid)
+add_subdirectory(07_local_mesh)
