### 2022-03-09 bb1b30e
Added string label argument to silo particle writer to support multiple particle groups.
