+++ cajita/unit_test/tstArray3d.hpp
@@ -36,7 +36,7 @@ void layoutTest()
     // Let MPI compute the partitioning for this test.
     DimBlockPartitioner<3> partitioner;
 
-    // Create the global .
+    // Create the global mesh.
     double cell_size = 0.23;
     std::array<int, 3> global_num_cell = { 37, 15, 20 };
     std::array<bool, 3> is_dim_periodic = { true, true, true };
