+++ example/cajita_tutorial/CMakeLists.txt
@@ -16,3 +16,4 @@ add_subdirectory(04_global_grid)
 add_subdirectory(05_index_space)
 add_subdirectory(06_local_grid)
 add_subdirectory(07_local_mesh)
+add_subdirectory(08_array)
