+++ example/cajita_tutorial/13_all_loadbalancer/loadbalancer_example.cpp
@@ -1,0 +1,158 @@
+/****************************************************************************
+ * Copyright (c) 2018-2022 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+#include <Cajita_LoadBalancer.hpp>
+
+#include <cmath>
+#include <iostream>
+
+//---------------------------------------------------------------------------//
+// A couple of helper functions.
+//---------------------------------------------------------------------------//
+// Creates an artificial work for a rectangular domain according to a Gaussian
+// work density.
+double workGauss( const std::array<double, 4>& verts,
+                  std::array<double, 2>& global_low_corner,
+                  std::array<double, 2>& global_high_corner )
+{
+    // The work density function is a radial gauss curve centred at (x0,y0)
+    // work density function: r = sqrt((x-x0)^2+(y-y0)^2)
+    //                        wd = 1/(sigma sqrt(2Pi)) Exp(-.5*( r/sigma )^2 )
+    // -.5*sqrt(Pi/2) sigma (
+    //                        Erf( (x1-x0)/(sqrt(2) sigma) ) -
+    //                        Erf( (x2-x0)/(sqrt(2) sigma) )
+    //                      ) (
+    //                        Erf( (y1-y0)/(sqrt(2) sigma) ) -
+    //                        Erf( (y2-y0)/(sqrt(2) sigma) )
+    //                      )
+    const double x1 = verts[0];
+    const double y1 = verts[1];
+    const double x2 = verts[2];
+    const double y2 = verts[3];
+    const double x0 = ( global_low_corner[0] + global_high_corner[0] ) / 2.;
+    const double y0 = ( global_low_corner[1] + global_high_corner[1] ) / 2.;
+    const double sigma = 1.;
+    const double s2sigma = 1. / ( sqrt( 2. ) * sigma );
+    const double Erfx1 = erf( ( x1 - x0 ) * s2sigma );
+    const double Erfx2 = erf( ( x2 - x0 ) * s2sigma );
+    const double Erfy1 = erf( ( y0 - y1 ) * s2sigma );
+    const double Erfy2 = erf( ( y0 - y2 ) * s2sigma );
+    return -0.5 * sqrt( 0.5 * M_PI ) * sigma * ( Erfx1 - Erfx2 ) *
+           ( Erfy1 - Erfy2 );
+}
+//---------------------------------------------------------------------------//
+// Loadbalancer example.
+//---------------------------------------------------------------------------//
+void loadBalancerExample()
+{
+    /*
+     * The loadbalancer class generates a new GlobalGrid given the local work.
+     * Most is the same as in any other program using Cajita. The comments will
+     * be focused on the additions/changes due to including the loadbalancer.
+     */
+    /*
+     * The example system is 2D and its size based on the number of ranks. Every
+     * rank will have the same initial 47 x 38 domain. The cell size is chosen
+     * at arbitrarily as is the offset of the system (global_low_corner). The
+     * system is non periodic. These choices are arbitrary with regards to the
+     * loadbalancer and were chosen for ease of use/consistency for different
+     * number of ranks.
+     *
+     * The partitioner, global mesh and (initial) global grid are set up as
+     * usual.
+     */
+    const Cajita::DimBlockPartitioner<2> partitioner;
+
+    std::array<int, 2> empty_array = { 0 };
+    std::array<int, 2> ranks =
[…]
