### 2022-03-13 233fe01
fixup: make sure Kokkos is finalized before MPI in Cajita examples
