+++ example/cajita_tutorial/03_partitioner/partitioner_example.cpp
@@ -96,11 +96,11 @@ void partitionerExample()
 int main( int argc, char* argv[] )
 {
     MPI_Init( &argc, &argv );
+    {
+        Kokkos::ScopeGuard scope_guard( argc, argv );
 
-    Kokkos::ScopeGuard scope_guard( argc, argv );
-
-    partitionerExample();
-
+        partitionerExample();
+    }
     MPI_Finalize();
 
     return 0;
