+++ example/cajita_tutorial/04_global_grid/global_grid_example.cpp
@@ -149,11 +149,11 @@ void globalGridExample()
 int main( int argc, char* argv[] )
 {
     MPI_Init( &argc, &argv );
+    {
+        Kokkos::ScopeGuard scope_guard( argc, argv );
 
-    Kokkos::ScopeGuard scope_guard( argc, argv );
-
-    globalGridExample();
-
+        globalGridExample();
+    }
     MPI_Finalize();
 
     return 0;
