+++ doxygen/search/functions_9.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['join_1194',['join',['../structCajita_1_1ArrayOp_1_1DotFunctor.html#a8e350c960292e28f719a2733e3b56704',1,'Cajita::ArrayOp::DotFunctor::join()'],['../structCajita_1_1ArrayOp_1_1NormInfFunctor.html#a3bf2bfc1b6c9f158df81aa3dfb390122',1,'Cajita::ArrayOp::NormInfFunctor::join()'],['../structCajita_1_1ArrayOp_1_1Norm1Functor.html#af09fd27b4e6c0f799de8398c5eecadb4',1,'Cajita::ArrayOp::Norm1Functor::join()'],['../structCajita_1_1ArrayOp_1_1Norm2Functor.html#a2bbd3b285ae7f7bf458df27f55b30f4a',1,'Cajita::ArrayOp::Norm2Functor::join()']]]
+  ['join_1200',['join',['../structCajita_1_1ArrayOp_1_1DotFunctor.html#a8e350c960292e28f719a2733e3b56704',1,'Cajita::ArrayOp::DotFunctor::join()'],['../structCajita_1_1ArrayOp_1_1NormInfFunctor.html#a3bf2bfc1b6c9f158df81aa3dfb390122',1,'Cajita::ArrayOp::NormInfFunctor::join()'],['../structCajita_1_1ArrayOp_1_1Norm1Functor.html#af09fd27b4e6c0f799de8398c5eecadb4',1,'Cajita::ArrayOp::Norm1Functor::join()'],['../structCajita_1_1ArrayOp_1_1Norm2Functor.html#a2bbd3b285ae7f7bf458df27f55b30f4a',1,'Cajita::ArrayOp::Norm2Functor::join()']]]
 ];
