+++ doxygen/search/all_15.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['uniformdimpartitioner_674',['UniformDimPartitioner',['../classCajita_1_1UniformDimPartitioner.html',1,'Cajita']]],
-  ['uniformmesh_675',['UniformMesh',['../structCajita_1_1UniformMesh.html',1,'Cajita']]],
-  ['unpackarray_676',['unpackArray',['../classCajita_1_1Halo.html#a39d1ce6406957eec5c46cbe8fc1201e3',1,'Cajita::Halo::unpackArray(const ReduceOp &amp;reduce_op, const Kokkos::View&lt; char *, memory_space &gt; &amp;buffer, const Kokkos::View&lt; int **, memory_space &gt; &amp;steering, const int element_idx, const std::integral_constant&lt; std::size_t, 0 &gt;, const Cabana::ParameterPack&lt; ArrayViews... &gt; &amp;array_views)'],['../classCajita_1_1Halo.html#a15adbfc0390a1aa9cf6746683c19915a',1,'Cajita::Halo::unpackArray(const ReduceOp reduce_op, const Kokkos::View&lt; char *, memory_space &gt; &amp;buffer, const Kokkos::View&lt; int **, memory_space &gt; &amp;steering, const int element_idx, const std::integral_constant&lt; std::size_t, N &gt;, const Cabana::ParameterPack&lt; ArrayViews... &gt; &amp;array_views)']]],
-  ['unpackbuffer_677',['unpackBuffer',['../classCajita_1_1Halo.html#acd3b468ef408354a4fa5db8ab747cf6c',1,'Cajita::Halo']]],
-  ['unpackelement_678',['unpackElement',['../classCajita_1_1Halo.html#a78912547aef7135f523e8c07c12af25f',1,'Cajita::Halo::unpackElement(const ReduceOp &amp;reduce_op, const Kokkos::View&lt; char *, memory_space &gt; &amp;buffer, const Kokkos::View&lt; int **, memory_space &gt; &amp;steering, const int element_idx, const ArrayView &amp;array_view)'],['../classCajita_1_1Halo.html#ab5d13bc61083e590de0bfcb0feb14b9e',1,'Cajita::Halo::unpackElement(const ReduceOp &amp;reduce_op, const Kokkos::View&lt; char *, memory_space &gt; &amp;buffer, const Kokkos::View&lt; int **, memory_space &gt; &amp;steering, const int element_idx, const ArrayView &amp;array_view)']]],
-  ['unpackop_679',['unpackOp',['../classCajita_1_1Halo.html#a630589f89353465615b73a28a5549f38',1,'Cajita::Halo::unpackOp(ScatterReduce::Sum, const T &amp;buffer_val, T &amp;array_val)'],['../classCajita_1_1Halo.html#a8877f8cf9a2d2392f61ecd946bb31adf',1,'Cajita::Halo::unpackOp(ScatterReduce::Min, const T &amp;buffer_val, T &amp;array_val)'],['../classCajita_1_1Halo.html#a2e2a956357264e4d4aada14abe60b3c6',1,'Cajita::Halo::unpackOp(ScatterReduce::Max, const T &amp;buffer_val, T &amp;array_val)'],['../classCajita_1_1Halo.html#aa5fb6d6201e1560cb2b3718d25ba0b2f',1,'Cajita::Halo::unpackOp(ScatterReduce::Replace, const T &amp;buffer_val, T &amp;array_val)']]],
[…]
