+++ doxygen/search/typedefs_c.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['random_5faccess_5fslice_1470',['random_access_slice',['../classCabana_1_1Slice.html#a3ceec95532b42f90d6ca1487a349e7ea',1,'Cabana::Slice']]],
-  ['reference_5ftype_1471',['reference_type',['../classCabana_1_1Slice.html#a889acbd8e5feb213e163c762057ff401',1,'Cabana::Slice']]]
+  ['random_5faccess_5fslice_1476',['random_access_slice',['../classCabana_1_1Slice.html#a3ceec95532b42f90d6ca1487a349e7ea',1,'Cabana::Slice']]],
+  ['reference_5ftype_1477',['reference_type',['../classCabana_1_1Slice.html#a889acbd8e5feb213e163c762057ff401',1,'Cabana::Slice']]]
 ];
