+++ doxygen/search/classes_4.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['edge_751',['Edge',['../structCajita_1_1Edge.html',1,'Cajita']]],
-  ['edge_3c_20dim_3a_3ai_20_3e_752',['Edge&lt; Dim::I &gt;',['../structCajita_1_1Edge_3_01Dim_1_1I_01_4.html',1,'Cajita']]],
-  ['edge_3c_20dim_3a_3aj_20_3e_753',['Edge&lt; Dim::J &gt;',['../structCajita_1_1Edge_3_01Dim_1_1J_01_4.html',1,'Cajita']]],
-  ['edge_3c_20dim_3a_3ak_20_3e_754',['Edge&lt; Dim::K &gt;',['../structCajita_1_1Edge_3_01Dim_1_1K_01_4.html',1,'Cajita']]]
+  ['edge_754',['Edge',['../structCajita_1_1Edge.html',1,'Cajita']]],
+  ['edge_3c_20dim_3a_3ai_20_3e_755',['Edge&lt; Dim::I &gt;',['../structCajita_1_1Edge_3_01Dim_1_1I_01_4.html',1,'Cajita']]],
+  ['edge_3c_20dim_3a_3aj_20_3e_756',['Edge&lt; Dim::J &gt;',['../structCajita_1_1Edge_3_01Dim_1_1J_01_4.html',1,'Cajita']]],
+  ['edge_3c_20dim_3a_3ak_20_3e_757',['Edge&lt; Dim::K &gt;',['../structCajita_1_1Edge_3_01Dim_1_1K_01_4.html',1,'Cajita']]]
 ];

[…]
