+++ example/cajita_tutorial/CMakeLists.txt
@@ -18,5 +18,5 @@ add_subdirectory(06_local_grid)
 add_subdirectory(07_local_mesh)
 add_subdirectory(08_array)
 if(Cabana_ENABLE_HEFFTE)
-    add_subdirectory(05_fast_fourier_transform)
+    add_subdirectory(10_fft_heffte)
 endif()
