### 2022-03-17 38a05b0
Update FFT example - previous version missing clone; (mirror did not create new view); other comment cleanup
