### 2022-03-17 8961ab8
fixup: spline and interpolation examples
