+++ example/cajita_tutorial/08_array/array_example.cpp
@@ -23,9 +23,10 @@
 void arrayExample()
 {
     /*
-      The array and corresponding array layout allocates and stores the physical
-      quantities on the grid - prior to creating an Array, only the size, shape,
-      and indexing of the grid is defined.
+      An array layout includes the local grid size information together with
+      the number of values per mesh entity in the field (dimensionality), and
+      the entity type on which the field is defined. An array stores the field
+      data itself and may only be defined on a single entity type.
     */
 
     int comm_rank = -1;
@@ -76,39 +77,43 @@ void arrayExample()
     int dofs_per_node = 4;
     auto node_layout = Cajita::createArrayLayout(
         global_grid, halo_width, dofs_per_node, Cajita::Node() );
+    int view_rank = node_layout->num_space_dim + 1;
 
     /*
       Similar to the local grid, the array layout holds index spaces for
       parallel iteration over mesh fields. These again include options for mesh
       entities and owned/owned+ghosted.
+
+      Note that index spaces generated by array layouts have an additional
+      dimension over the field values on each entity.
     */
     auto array_node_owned_space =
         node_layout->indexSpace( Cajita::Own(), Cajita::Local() );
     std::cout << "Array layout (Own, Local) \nMin: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_owned_space.min( d ) << " ";
     std::cout << "\nMax: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_owned_space.max( d ) << " ";
     std::cout << "\n" << std::endl;
 
     auto array_node_ghosted_space =
         node_layout->indexSpace( Cajita::Ghost(), Cajita::Local() );
     std::cout << "Array layout (Ghost, Local) \nMin: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_ghosted_space.min( d ) << " ";
     std::cout << "\nMax: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_ghosted_space.max( d ) << " ";
     std::cout << "\n" << std::endl;
 
     auto array_node_shared_owned_space =
         node_layout->sharedIndexSpace( Cajita::Own(), -1, 0, 1 );
     std::cout << "Array layout (Shared, Own, -1,0,1) \nMin: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_shared_owned_space.min( d ) << " ";
     std::cout << "\nMax: ";
-    for ( int d = 0; d < dofs_per_node; ++d )
+    for ( int d = 0; d < view_rank; ++d )
         std::cout << array_node_shared_owned_space.max( d ) << " ";
     std::cout << "\n" << std::endl;
 
@@ -123,7 +128,7 @@ void arrayExample()
     auto view = array->view();
     std::cout << "Array total size: " << view.size() << std::endl;
     std::cout << "Array extents: ";
-    for ( int i = 0; i < 4; ++i )
+    for ( int i = 0; i < view_rank; ++i )
         std::cout << view.extent( i ) << " ";
     std::cout << "\n" << std::endl;
 
@@ -145,7 +150,12 @@ void arrayExample()
     // Scale the entire array with a single value.
[…]
