### 2022-03-24 ada2968
fix gtest not found issue
