+++ cmake/test_harness/test_harness.cmake
@@ -24,7 +24,7 @@ macro(Cabana_add_tests_nobackend)
   foreach(_test ${CABANA_UNIT_TEST_NAMES})
     set(_target Cabana_${_test}_test)
     add_executable(${_target} tst${_test}.cpp ${TEST_HARNESS_DIR}/unit_test_main.cpp)
-    target_link_libraries(${_target} PRIVATE ${CABANA_UNIT_TEST_PACKAGE} gtest)
+    target_link_libraries(${_target} PRIVATE ${CABANA_UNIT_TEST_PACKAGE} GTest::gtest)
     add_test(NAME ${_target} COMMAND ${NONMPI_PRECOMMAND} ${_target} ${gtest_args})
     set_property(TEST ${_target} PROPERTY ENVIRONMENT OMP_NUM_THREADS=1)
     if(VALGRIND_FOUND)
@@ -86,7 +86,7 @@ macro(Cabana_add_tests)
       add_executable(${_target} ${_file} ${CABANA_UNIT_TEST_MAIN})
       target_include_directories(${_target} PRIVATE ${_dir}
         ${TEST_HARNESS_DIR} ${CMAKE_CURRENT_SOURCE_DIR})
-      target_link_libraries(${_target} PRIVATE ${CABANA_UNIT_TEST_PACKAGE} gtest)
+      target_link_libraries(${_target} PRIVATE ${CABANA_UNIT_TEST_PACKAGE} GTest::gtest)
       if(CABANA_UNIT_TEST_MPI)
         foreach(_np ${CABANA_UNIT_TEST_MPIEXEC_NUMPROCS})
           add_test(NAME ${_target}_np_${_np} COMMAND
