### 2022-03-25 cd87146
Remove gtest auto-download and internal build - treat as normal dependency
