### 2022-03-28 #4115 @streeve
@diehlpk since we have relatively new developers adding our new examples and performance analysis I wanted to make sure it was okay to add them to the author list now
