### 2022-03-29 6946755
Add gtest to jenkins builds
