+++ docker/Dockerfile
@@ -8,6 +8,7 @@ RUN apt-get update && apt-get install -y \
         ccache \
         wget \
         openssh-client \
+        libgtest-dev \
     && \
     apt-get clean && \
     rm -rf /var/lib/apt/lists/*
