+++ benchmark/core/Cabana_NeighborArborXPerformance.cpp
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: BSD-3-Clause                                    *
  ****************************************************************************/
 
-#include "Cabana_BenchmarkUtils.hpp"
+#include "../Cabana_BenchmarkUtils.hpp"
 
 #include <Cabana_Core.hpp>
 
@@ -28,8 +28,15 @@
 template <class Device>
 void performanceTest( std::ostream& stream, const std::string& test_prefix,
                       std::vector<int> problem_sizes,
-                      std::vector<double> cutoff_ratios )
+                      std::vector<double> cutoff_ratios, bool sort = true )
 {
+    using exec_space = typename Device::execution_space;
+    using memory_space = typename Device::memory_space;
+
+    // Declare the neighbor list type.
+    using ListTag = Cabana::FullNeighborTag;
+    using IterTag = Cabana::SerialOpTag;
+
     // Declare problem sizes.
     double min_dist = 1.0;
     int num_problem_size = problem_sizes.size();
@@ -60,6 +67,22 @@ void performanceTest( std::ostream& stream, const std::string& test_prefix,
         auto x = Cabana::slice<0>( aosoas[p], "position" );
         Cabana::createRandomParticlesMinDistance( x, x.size(), x_min[p],
                                                   x_max[p], min_dist );
+
+        if ( sort )
+        {
+            // Sort the particles to make them more realistic, e.g. in an MD
+            // simulation. They likely won't be randomly scattered about, but
+            // rather will be periodically sorted for spatial locality. Bin them
+            // in cells the size of the smallest cutoff distance.
+            double cutoff = cutoff_ratios.front() * min_dist;
+            double sort_delta[3] = { cutoff, cutoff, cutoff };
+            double grid_min[3] = { x_min[p], x_min[p], x_min[p] };
+            double grid_max[3] = { x_max[p], x_max[p], x_max[p] };
+            auto x = Cabana::slice<0>( aosoas[p], "position" );
+            Cabana::LinkedCellList<Device> linked_cell_list(
+                x, sort_delta, grid_min, grid_max );
+            Cabana::permute( linked_cell_list, aosoas[p] );
+        }
     }
 
     // Loop over number of ratios (neighbors per particle).
@@ -69,16 +92,17 @@ void performanceTest( std::ostream& stream, const std::string& test_prefix,
 
         // Create timers.
         std::stringstream create_time_name;
-        create_time_name << test_prefix << "linkedcell_create_"
-                         << cutoff_ratios[c];
+        create_time_name << test_prefix << "neigh_create_" << cutoff_ratios[c];
         Cabana::Benchmark::Timer create_timer( create_time_name.str(),
                                                num_problem_size );
-        std::stringstream sort_time_name;
-        sort_time_name << test_prefix << "linkedcell_sort_" << cutoff_ratios[c];
-        Cabana::Benchmark::Timer sort_timer( sort_time_name.str(),
-                                             num_problem_size );
+        std::stringstream iteration_time_name;
+        iteration_time_name << test_prefix << "neigh_iteration_"
+                            << cutoff_ratios[c];
+        Cabana::Benchmark::Timer iteration_timer( iteration_time_name.str(),
+                                                  num_problem_size );
 
         // Loop over the problem sizes.
+        int pid = 0;
         std::vector<int> psizes;
         for ( int p = 0; p < num_problem_size; ++p )
         {
@@ -89,33 +113,43 @@ void performanceTest( std::ostream& stream, const std::string& test_prefix,
             // Track the problem size.
             psizes.push_back( problem_sizes[p] );
 
-            // Create the linked cell list.
-            auto x = Cabana::slice<0>( aosoas[p], "position" );
[…]
