+++ benchmark/CMakeLists.txt
@@ -9,70 +9,8 @@
 # SPDX-License-Identifier: BSD-3-Clause                                    #
 ############################################################################
 
-if(Kokkos_ENABLE_SERIAL OR Kokkos_ENABLE_CUDA OR Kokkos_ENABLE_OPENMP OR Kokkos_ENABLE_HIP)
-  add_executable(BinSortPerformance Cabana_BinSortPerformance.cpp)
-  target_link_libraries(BinSortPerformance cabanacore)
+add_subdirectory(core)
 
-  add_executable(NeighborVerletPerformance Cabana_NeighborVerletPerformance.cpp)
-  target_link_libraries(NeighborVerletPerformance cabanacore)
-
-  if(Cabana_ENABLE_ARBORX)
-    add_executable(NeighborArborXPerformance Cabana_NeighborArborXPerformance.cpp)
-    target_link_libraries(NeighborArborXPerformance cabanacore)
-  endif()
-
-  add_executable(LinkedCellPerformance Cabana_LinkedCellPerformance.cpp)
-  target_link_libraries(LinkedCellPerformance cabanacore)
-
-  if(Cabana_ENABLE_MPI)
-    add_executable(CommPerformance Cabana_CommPerformance.cpp)
-    target_link_libraries(CommPerformance cabanacore)
-  endif()
-
-  if(Cabana_ENABLE_CAJITA)
-    add_executable(SparseMapPerformance Cajita_SparseMapPerformance.cpp)
-    target_link_libraries(SparseMapPerformance Cajita)
-
-    add_executable(SparsePartitionerPerformance Cajita_SparsePartitionerPerformance.cpp)
-    target_link_libraries(SparsePartitionerPerformance Cajita)
-
-    if(Cabana_ENABLE_HEFFTE)
-        add_executable(FastFourierTransformPerformance Cajita_FastFourierTransformPerformance.cpp)
-        target_link_libraries(FastFourierTransformPerformance Cajita)
-    endif()
-  endif()
-
-  if(Cabana_ENABLE_TESTING)
-    target_link_libraries(BinSortPerformance gtest)
-    add_test(NAME Cabana_Performance_BinSort COMMAND ${NONMPI_PRECOMMAND} BinSortPerformance sort_output.txt)
-
-    target_link_libraries(NeighborVerletPerformance gtest)
-    add_test(NAME Cabana_Performance_NeighborVerlet COMMAND ${NONMPI_PRECOMMAND} NeighborVerletPerformance verlet_output.txt)
-
-    if(Cabana_ENABLE_ARBORX)
-      target_link_libraries(NeighborArborXPerformance gtest)
-      add_test(NAME Cabana_Performance_NeighborArborX COMMAND ${NONMPI_PRECOMMAND} NeighborArborXPerformance arborx_output.txt)
-    endif()
-
-    target_link_libraries(LinkedCellPerformance gtest)
-    add_test(NAME Cabana_Performance_LinkedCell COMMAND ${NONMPI_PRECOMMAND} LinkedCellPerformance lcl_output.txt)
-
-    if(Cabana_ENABLE_MPI)
-      target_link_libraries(CommPerformance gtest)
-      add_test(NAME Cabana_Performance_Comm COMMAND ${NONMPI_PRECOMMAND} CommPerformance 1000 comm_output.txt)
-    endif()
-
-    if (Cabana_ENABLE_CAJITA)
-      target_link_libraries(SparseMapPerformance gtest)
-      add_test(NAME Cajita_SparseMapPerformance COMMAND ${NONMPI_PRECOMMAND} SparseMapPerformance sparsemap_output.txt)
-
[…]
