### 2022-03-29 #4115 @streeve
From our side, new examples are nearly finished (ECP-copa/Cabana/issues/352 & ECP-copa/Cabana/issues/496). Wiki is nearing a complete draft for Cajita (https://github.com/ECP-copa/Cabana/wiki/Programming-Guide) and performance (https://github.com/ECP-copa/Cabana/wiki/Benchmarks)
