+++ example/cajita_tutorial/12_halo/halo_example.cpp
@@ -1,0 +1,172 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include <mpi.h>
+
+#include <array>
+
+//---------------------------------------------------------------------------//
+void gatherScatterExample( const Cajita::ManualBlockPartitioner<3>& partitioner,
+                           const std::array<bool, 3>& is_dim_periodic,
+                           const std::string test_name )
+{
+    /*
+      The halo is a communication plan designed from halo exchange where some
+      locally-owned elements on each rank are used as ghost data on other
+      ranks. The halo supplies both forward and reverse communication
+      operations. In the forward operation (the gather), data is sent from the
+      uniquely-owned decomposition to the ghosted decomposition. In the
+      reverse operation (the scatter), data is sent from the ghosted
+      decomposition back to the uniquely-owned decomposition and collisions
+      are resolved.
+
+      In this example we will demonstrate building a halo communication
+      plan and performing both scatter and gather operations.
+    */
+
+    int comm_rank = -1;
+    MPI_Comm_rank( MPI_COMM_WORLD, &comm_rank );
+
+    if ( comm_rank == 0 )
+    {
+        std::cout << "=======================\n";
+        std::cout << test_name << std::endl;
+        std::cout << "=======================\n";
+    }
+
+    using exec_space = Kokkos::DefaultHostExecutionSpace;
+    using device_type = exec_space::device_type;
+
+    // Create the global grid.
+    double cell_size = 0.23;
+    std::array<int, 3> global_num_cell = { 16, 21, 21 };
+    std::array<double, 3> global_low_corner = { 1.2, 3.3, -2.8 };
+    std::array<double, 3> global_high_corner = {
+        global_low_corner[0] + cell_size * global_num_cell[0],
+        global_low_corner[1] + cell_size * global_num_cell[1],
+        global_low_corner[2] + cell_size * global_num_cell[2] };
+    auto global_mesh = Cajita::createUniformGlobalMesh(
+        global_low_corner, global_high_corner, global_num_cell );
+
+    // Create the global grid.
+    auto global_grid = createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                         is_dim_periodic, partitioner );
+
+    // Array parameters.
+    unsigned array_halo_width = 3;
+
+    // Loop over halo sizes up to the size of the array halo width.
+    for ( unsigned halo_width = 1; halo_width <= array_halo_width;
+          ++halo_width )
+    {
+        if ( comm_rank == 0 )
+            std::cout << "halo width : " << halo_width << std::endl;
+
+        // Create a cell array.
+        auto layout = createArrayLayout( global_grid, array_halo_width, 4,
+                                         Cajita::Cell() );
+        auto array =
+            Cajita::createArray<double, device_type>( "array", layout );
+
+        // Assign the owned cells a value of 1 and the rest 0.
+        Cajita::ArrayOp::assign( *array, 0.0, Cajita::Ghost() );
[…]
