### 2022-03-31 dafcdb9
Update grid halo example
