+++ example/cajita_tutorial/12_halo/halo_example.cpp
@@ -18,9 +18,9 @@
 #include <array>
 
 //---------------------------------------------------------------------------//
-void gatherScatterExample( const Cajita::ManualBlockPartitioner<3>& partitioner,
-                           const std::array<bool, 3>& is_dim_periodic,
-                           const std::string test_name )
+// Grid Halo example.
+//---------------------------------------------------------------------------//
+void gridHaloExample()
 {
     /*
       The halo is a communication plan designed from halo exchange where some
@@ -30,7 +30,8 @@ void gatherScatterExample( const Cajita::ManualBlockPartitioner<3>& partitioner,
       uniquely-owned decomposition to the ghosted decomposition. In the
       reverse operation (the scatter), data is sent from the ghosted
       decomposition back to the uniquely-owned decomposition and collisions
-      are resolved.
+      are resolved. Grid halos in Cajita are relvatively simple because all
+      grids are logically rectilinear.
 
       In this example we will demonstrate building a halo communication
       plan and performing both scatter and gather operations.
@@ -41,17 +42,35 @@ void gatherScatterExample( const Cajita::ManualBlockPartitioner<3>& partitioner,
 
     if ( comm_rank == 0 )
     {
-        std::cout << "=======================\n";
-        std::cout << test_name << std::endl;
-        std::cout << "=======================\n";
+        std::cout << "Cajita Grid Halo Example" << std::endl;
+        std::cout << "    (intended to be run with MPI)\n" << std::endl;
     }
 
     using exec_space = Kokkos::DefaultHostExecutionSpace;
     using device_type = exec_space::device_type;
 
-    // Create the global grid.
+    // Use linear MPI partitioning to make this example simpler.
+    int comm_size;
+    MPI_Comm_size( MPI_COMM_WORLD, &comm_size );
+    std::array<int, 3> ranks_per_dim = { comm_size, 1, 1 };
+    MPI_Dims_create( comm_size, 3, ranks_per_dim.data() );
+    Cajita::ManualBlockPartitioner<3> partitioner( ranks_per_dim );
+
+    /*
+      Boundaries are not periodic in this example, thus ranks at system
+      boundaries do not communicate. Because of this the details printed from
+      rank 0 on the left side will never change as it is on the system boundary
+      (without an MPI neighbor).
+
+      Changing values here to true will show what happens for periodic grid
+      communication instead, where the left and right sides of the local domain
+      should then match.
+    */
+    std::array<bool, 3> is_dim_periodic = { false, false, false };
+
+    // Create the mesh and grid structures as usual.
     double cell_size = 0.23;
-    std::array<int, 3> global_num_cell = { 16, 21, 21 };
+    std::array<int, 3> global_num_cell = { 26, 21, 21 };
     std::array<double, 3> global_low_corner = { 1.2, 3.3, -2.8 };
     std::array<double, 3> global_high_corner = {
         global_low_corner[0] + cell_size * global_num_cell[0],
@@ -59,79 +78,101 @@ void gatherScatterExample( const Cajita::ManualBlockPartitioner<3>& partitioner,
         global_low_corner[2] + cell_size * global_num_cell[2] };
     auto global_mesh = Cajita::createUniformGlobalMesh(
         global_low_corner, global_high_corner, global_num_cell );
-
-    // Create the global grid.
     auto global_grid = createGlobalGrid( MPI_COMM_WORLD, global_mesh,
                                          is_dim_periodic, partitioner );
 
-    // Array parameters.
-    unsigned array_halo_width = 3;
+    /*
+      Here the halo width allocated for the system is not necessarily always
+      fully communicated - we can use any integer value from zero to the value
+      allocated.
+    */
+    unsigned allocated_halo_width = 3;
 
[…]
