### 2022-03-31 bb7b307
Address review comments and expound usage scenarios.
