+++ core/performance_test/peakflops/01_cpp_simple.cpp
@@ -1,9 +1,21 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
 #include "common.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
 #include <gtest/gtest.h>
+
 /*
 For skylake
 Model name:            Intel(R) Xeon(R) Gold 6152 CPU @ 2.10GHz
