+++ example/tutorial/08_linked_cell_list/linked_cell_list_example.cpp
@@ -1,147 +1,0 @@
-/****************************************************************************
- * Copyright (c) 2018-2020 by the Cabana authors                            *
- * All rights reserved.                                                     *
- *                                                                          *
- * This file is part of the Cabana library. Cabana is distributed under a   *
- * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
- * the top-level directory.                                                 *
- *                                                                          *
- * SPDX-License-Identifier: BSD-3-Clause                                    *
- ****************************************************************************/
-
-#include <Cabana_Core.hpp>
-
-#include <iostream>
-
-//---------------------------------------------------------------------------//
-// Linked cell list.
-//---------------------------------------------------------------------------//
-void linkedCellListExample()
-{
-    /*
-      A linked cell list is a data structure which bins particles based on the
-      Cartesian grid cell in which they are located. An input uniform
-      Cartesian grid description is used to locate which particles, defined by
-      an input slice of 3D Cartesian coordinates, are located in which grid
-      cells.
-
-      The result of creating this data structure is binning data which may
-      then be used to reorder the data in an AoSoA in identical fashion to the
-      sorting and binning example we just completed. In effect, a linked cell
-      list is simply a more sophisticated form of binning which uses a custom
-      comparator for particles based on their spatial location relative to a
-      uniform Cartesian grid.
-    */
-
-    /*
-       Start by declaring the types in our tuples will store. The first
-       member will be the coordinates, the second an id.
-    */
-    using DataTypes = Cabana::MemberTypes<double[3], int>;
-
-    /*
-      Next declare the data layout of the AoSoA. We use the host space here
-      for the purposes of this example but all memory spaces, vector lengths,
-      and member type configurations are compatible with sorting.
-    */
-    const int VectorLength = 8;
-    using MemorySpace = Kokkos::HostSpace;
-    using ExecutionSpace = Kokkos::Serial;
-    using DeviceType = Kokkos::Device<ExecutionSpace, MemorySpace>;
-
-    /*
-       Create the AoSoA.
-    */
-    int num_tuple = 54;
-    Cabana::AoSoA<DataTypes, DeviceType, VectorLength> aosoa( "A", num_tuple );
-
-    /*
-      Define the parameters of the Cartesian grid over which we will build the
-      cell list. This is a simple 3x3x3 uniform grid on [0,3] in each
-      direction. Each grid cell has a size of 1 in each dimension.
-     */
-    double grid_min[3] = { 0.0, 0.0, 0.0 };
-    double grid_max[3] = { 3.0, 3.0, 3.0 };
-    double grid_delta[3] = { 1.0, 1.0, 1.0 };
-
-    /*
-      Create the particle ids.
-    */
-    auto ids = Cabana::slice<1>( aosoa );
-    for ( std::size_t i = 0; i < aosoa.size(); ++i )
-        ids( i ) = i;
-
-    /*
-      Create the particle coordinates. We will put 2 particles in the center
-      of each cell. We are ordering this such that each consecutive particle
-      is in a different cell. When we use cell list to permute the particles
-      later in the example, they will be regrouped by cell.
-    */
-    auto positions = Cabana::slice<0>( aosoa );
-    int ppc = 2;
-    int particle_counter = 0;
-    for ( int p = 0; p < ppc; ++p )
-        for ( int i = 0; i < 3; ++i )
-            for ( int j = 0; j < 3; ++j )
-                for ( int k = 0; k < 3; ++k, ++particle_counter )
-                {
-                    positions( particle_counter, 0 ) =
-                        grid_min[0] + grid_delta[0] * ( 0.5 + i );
-                    positions( particle_counter, 1 ) =
-                        grid_min[1] + grid_delta[1] * ( 0.5 + j );
[…]
