### 2022-03-31 99f613d
Removed unused typedef
