### 2022-03-31 #4115 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1177/10943420211022829 is OK
> - 10.1016/j.jpdc.2014.07.003 is OK
> - 10.1109/TPDS.2021.3097283 is OK
> - 10.21105/joss.01370 is OK
> - 10.1098/rsta.2019.0056 is OK
> - 10.1145/3412558 is OK
> - 10.1109/ExaMPI49596.2019.00007 is OK
> - 10.1007/3-540-47789-6_66 is OK
> - 10.1016/j.newast.2015.06.003 is OK
> - 10.1006/jcph.1995.1039 is OK
> - 10.1063/1.5020792 is OK
>
> MISSING DOIs
>
> - 10.1016/j.cpc.2021.108262 may be a valid DOI for title: N ways to simulate short-range particle systems: Automated algorithm selection with the node-level library AutoPas
> - 10.1016/j.cpc.2019.03.007 may be a valid DOI for title: OpenFPM: A scalable open framework for particle and particle-mesh codes on parallel computers
>
> INVALID DOIs
>
> - None
> ```

 @streeve could you please check the reported missing DOIs?
