### 2022-04-01 cf45928
Add city and country to affiliations
