### 2022-04-01 541dcd6
draft of structured solver example
