+++ example/cajita_tutorial/11_structured_solver/structured_solver_example.cpp
@@ -1,0 +1,212 @@
+/****************************************************************************
+ * Copyright (c) 2018-2021 by the Cabana authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the Cabana library. Cabana is distributed under a   *
+ * BSD 3-clause license. For the licensing terms see the LICENSE file in    *
+ * the top-level directory.                                                 *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#include <Cajita.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include <mpi.h>
+
+#include <array>
+#include <vector>
+
+//---------------------------------------------------------------------------//
+// Structured Solver Example
+//---------------------------------------------------------------------------//
+void structuredSolverExample()
+{
+
+    using MemorySpace = Kokkos::HostSpace;
+    using ExecutionSpace = Kokkos::DefaultHostExecutionSpace;
+
+    // Create the global grid.
+    double cell_size = 0.25;
+    std::array<bool, 3> is_dim_periodic = { false, false, false };
+    std::array<double, 3> global_low_corner = { -1.0, -2.0, -1.0 };
+    std::array<double, 3> global_high_corner = { 1.0, 1.0, 0.5 };
+    auto global_mesh = Cajita::createUniformGlobalMesh( global_low_corner,
+                                                global_high_corner, cell_size );
+
+    // Create the global grid.
+    Cajita::DimBlockPartitioner<3> partitioner;
+    auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
+                                         is_dim_periodic, partitioner );
+
+    // Create a local grid.
+    auto local_mesh = createLocalGrid( global_grid, 1 );
+    auto owned_space = local_mesh->indexSpace( Cajita::Own(), Cajita::Cell(), Cajita::Local() );
+
+    // Create the RHS.
+    auto vector_layout = createArrayLayout( local_mesh, 1, Cajita::Cell() );
+    auto rhs = Cajita::createArray<double, MemorySpace>( "rhs", vector_layout );
+    Cajita::ArrayOp::assign( *rhs, 1.0, Cajita::Own() );
+
+    // Create the LHS.
+    auto lhs = Cajita::createArray<double, MemorySpace>( "lhs", vector_layout );
+    Cajita::ArrayOp::assign( *lhs, 0.0, Cajita::Own() );
+
+    // Create a solver.
+    auto solver = Cajita::createHypreStructuredSolver<double, MemorySpace>(
+        "PCG", *vector_layout );
+
+    // Create a 7-point 3d laplacian stencil.
+    std::vector<std::array<int, 3>> stencil = {
+        { 0, 0, 0 }, { -1, 0, 0 }, { 1, 0, 0 }, { 0, -1, 0 },
+        { 0, 1, 0 }, { 0, 0, -1 }, { 0, 0, 1 } };
+    solver->setMatrixStencil( stencil );
+
+    // Create the matrix entries. The stencil is defined over cells.
+    auto matrix_entry_layout = createArrayLayout( local_mesh, 7, Cajita::Cell() );
+    auto matrix_entries = Cajita::createArray<double, MemorySpace>(
[…]
