+++ example/cajita_tutorial/11_structured_solver/structured_solver_example.cpp
@@ -32,17 +32,18 @@ void structuredSolverExample()
     std::array<bool, 3> is_dim_periodic = { false, false, false };
     std::array<double, 3> global_low_corner = { -1.0, -2.0, -1.0 };
     std::array<double, 3> global_high_corner = { 1.0, 1.0, 0.5 };
-    auto global_mesh = Cajita::createUniformGlobalMesh( global_low_corner,
-                                                global_high_corner, cell_size );
+    auto global_mesh = Cajita::createUniformGlobalMesh(
+        global_low_corner, global_high_corner, cell_size );
 
     // Create the global grid.
     Cajita::DimBlockPartitioner<3> partitioner;
     auto global_grid = Cajita::createGlobalGrid( MPI_COMM_WORLD, global_mesh,
-                                         is_dim_periodic, partitioner );
+                                                 is_dim_periodic, partitioner );
 
     // Create a local grid.
     auto local_mesh = createLocalGrid( global_grid, 1 );
-    auto owned_space = local_mesh->indexSpace( Cajita::Own(), Cajita::Cell(), Cajita::Local() );
+    auto owned_space = local_mesh->indexSpace( Cajita::Own(), Cajita::Cell(),
+                                               Cajita::Local() );
 
     // Create the RHS.
     auto vector_layout = createArrayLayout( local_mesh, 1, Cajita::Cell() );
@@ -64,7 +65,8 @@ void structuredSolverExample()
     solver->setMatrixStencil( stencil );
 
     // Create the matrix entries. The stencil is defined over cells.
-    auto matrix_entry_layout = createArrayLayout( local_mesh, 7, Cajita::Cell() );
+    auto matrix_entry_layout =
+        createArrayLayout( local_mesh, 7, Cajita::Cell() );
     auto matrix_entries = Cajita::createArray<double, MemorySpace>(
         "matrix_entries", matrix_entry_layout );
     auto entry_view = matrix_entries->view();
@@ -96,8 +98,9 @@ void structuredSolverExample()
     std::string precond_type = "Diagonal";
     if ( "none" != precond_type )
     {
-        auto preconditioner = Cajita::createHypreStructuredSolver<double, MemorySpace>(
-            precond_type, *vector_layout, true );
+        auto preconditioner =
+            Cajita::createHypreStructuredSolver<double, MemorySpace>(
+                precond_type, *vector_layout, true );
         solver->setPreconditioner( preconditioner );
     }
 
@@ -108,25 +111,34 @@ void structuredSolverExample()
     solver->solve( *rhs, *lhs );
 
     // Create a solver reference for comparison.
-    auto lhs_ref = Cajita::createArray<double, MemorySpace>( "lhs_ref", vector_layout );
+    auto lhs_ref =
+        Cajita::createArray<double, MemorySpace>( "lhs_ref", vector_layout );
     Cajita::ArrayOp::assign( *lhs_ref, 0.0, Cajita::Own() );
 
     auto ref_solver =
-        Cajita::createReferenceConjugateGradient<double, MemorySpace>( *vector_layout );
+        Cajita::createReferenceConjugateGradient<double, MemorySpace>(
+            *vector_layout );
     ref_solver->setMatrixStencil( stencil );
     const auto& ref_entries = ref_solver->getMatrixValues();
     auto matrix_view = ref_entries.view();
-    auto global_space = local_mesh->indexSpace( Cajita::Own(), Cajita::Cell(), Cajita::Global() );
[…]
