### 2022-04-01 22db17a
Separate reference and hypre solver examples
