+++ example/cajita_tutorial/11_structured_solver_hypre/hypre_structured_solver_example.cpp
@@ -19,11 +19,11 @@
 #include <vector>
 
 //---------------------------------------------------------------------------//
-// Structured Solver Example
+// HYPRE Structured Solver Example
 //---------------------------------------------------------------------------//
-void structuredSolverExample()
+void hypreStructuredSolverExample()
 {
-    std::cout << "Cajita Structured Solver Example\n" << std::endl;
+    std::cout << "Cajita HYPRE Structured Solver Example\n" << std::endl;
 
     using MemorySpace = Kokkos::HostSpace;
     using ExecutionSpace = Kokkos::DefaultHostExecutionSpace;
@@ -51,67 +51,71 @@ void structuredSolverExample()
     auto rhs = Cajita::createArray<double, MemorySpace>( "rhs", vector_layout );
     Cajita::ArrayOp::assign( *rhs, 1.0, Cajita::Own() );
 
+    // Create the LHS.
+    auto lhs = Cajita::createArray<double, MemorySpace>( "lhs", vector_layout );
+    Cajita::ArrayOp::assign( *lhs, 0.0, Cajita::Own() );
+
+    // Create a solver.
+    auto solver = Cajita::createHypreStructuredSolver<double, MemorySpace>(
+        "PCG", *vector_layout );
+
     // Create a 7-point 3d laplacian stencil.
     std::vector<std::array<int, 3>> stencil = {
         { 0, 0, 0 }, { -1, 0, 0 }, { 1, 0, 0 }, { 0, -1, 0 },
         { 0, 1, 0 }, { 0, 0, -1 }, { 0, 0, 1 } };
-    // Create a solver reference for comparison.
-    auto lhs_ref =
-        Cajita::createArray<double, MemorySpace>( "lhs_ref", vector_layout );
-    Cajita::ArrayOp::assign( *lhs_ref, 0.0, Cajita::Own() );
+    solver->setMatrixStencil( stencil );
 
-    auto ref_solver =
-        Cajita::createReferenceConjugateGradient<double, MemorySpace>(
-            *vector_layout );
-    ref_solver->setMatrixStencil( stencil );
-    const auto& ref_entries = ref_solver->getMatrixValues();
-    auto matrix_view = ref_entries.view();
-    auto global_space = local_mesh->indexSpace( Cajita::Own(), Cajita::Cell(),
-                                                Cajita::Global() );
-    int ncell_i =
-        global_grid->globalNumEntity( Cajita::Cell(), Cajita::Dim::I );
-    int ncell_j =
-        global_grid->globalNumEntity( Cajita::Cell(), Cajita::Dim::J );
-    int ncell_k =
-        global_grid->globalNumEntity( Cajita::Cell(), Cajita::Dim::K );
+    // Create the matrix entries. The stencil is defined over cells.
+    auto matrix_entry_layout =
+        createArrayLayout( local_mesh, 7, Cajita::Cell() );
+    auto matrix_entries = Cajita::createArray<double, MemorySpace>(
+        "matrix_entries", matrix_entry_layout );
+    auto entry_view = matrix_entries->view();
     Kokkos::parallel_for(
-        "fill_ref_entries",
+        "fill_matrix_entries",
         createExecutionPolicy( owned_space, ExecutionSpace() ),
         KOKKOS_LAMBDA( const int i, const int j, const int k ) {
-            int gi = i + global_space.min( Cajita::Dim::I ) -
-                     owned_space.min( Cajita::Dim::I );
-            int gj = j + global_space.min( Cajita::Dim::J ) -
[…]
