### 2022-04-01 15d139d
Disable on hypre example if device is enabled
