+++ example/cajita_tutorial/CMakeLists.txt
@@ -22,7 +22,7 @@ if(Cabana_ENABLE_HEFFTE)
     add_subdirectory(10_fft_heffte)
 endif()
 add_subdirectory(11_structured_solver)
-if(Cabana_ENABLE_HYPRE)
+if(Cabana_ENABLE_HYPRE AND (NOT Kokkos_ENABLE_CUDA AND NOT Kokkos_ENABLE_HIP AND NOT Kokkos_ENABLE_SYCL))
   add_subdirectory(11_structured_solver_hypre)
 endif()
 add_subdirectory(12_halo)
