### 2022-04-01 0a39341
Added thread-level interface example functors. More text.
