+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -14,6 +14,103 @@
 
 #include <iostream>
 
+/******************************************************************************
+ * User-defined thread-level functors may be used instead of the built-in
+ * functors provided. Here, the definition of P2GExampleFunctor directly uses
+ * the thread-level interpolation interface to perform some basic Cajita::P2G::
+ * interpolations.
+ * *****************************************************************************/
+template <class ScalarView, class VectorView>
+struct P2GExampleFunctor
+{
+    /* Value types. */
+    using scalar_value_type = typename ScalarView::value_type;
+    using vector_value_type = typename VectorView::value_type;
+
+    /* Spline evaluation locations. */
+    ScalarView _x;
+    VectorView _y;
+
+    /* Here, the constructor accepts both a scalar and a vector View of the
+     * particle data to interpolate
+     */
+    P2GExampleFunctor( const ScalarView& x, const VectorView& y )
+        : _x( x )
+        , _y( y )
+    {
+        static_assert( 1 == ScalarView::Rank, "First View must be of scalars" );
+        static_assert( 2 == VectorView::Rank,
+                       "Second View must be of vectors" );
+    }
+
+    /* Apply spline interpolation. */
+    template <class SplineDataType, class GridViewType>
+    KOKKOS_INLINE_FUNCTION void operator()( const SplineDataType& sd,
+                                            const int p,
+                                            const GridViewType& view ) const
+    {
+        /* Access the point data */
+        scalar_value_type scalar_point_data = _x( p );
+        vector_value_type vector_point_data[2];
+
+        /* Thread-local manipulations may be computed here. */
+        for ( int d = 0; d < 2; ++d )
+        {
+            vector_point_data[d] = 3.0 * _y( p, d );
+        }
+        scalar_point_data *= 2.0;
+
+        /* Finally, call the thread-level value spline interpolations. */
+        Cajita::P2G::value( scalar_point_data, sd, view );
+        Cajita::P2G::divergence( vector_point_data, sd, view );
+    }
+};
+
+template <class ScalarView, class TensorView>
+struct G2PExampleFunctor
+{
+    /* Value types. */
+    using scalar_value_type = typename ScalarView::value_type;
+    using tensor_value_type = typename TensorView::value_type;
+
+    /* Spline evaluation locations. */
+    ScalarView _x;
+    TensorView _t;
+
+    /* Here, the constructor accepts both a scalar and a tensor View of the
+     * particle data to interpolate
+     */
+    G2PExampleFunctor( const ScalarView& x, const TensorView& t )
+        : _x( x )
+        , _t( t )
+    {
+        static_assert( 1 == ScalarView::Rank, "First View must be of scalars" );
+        static_assert( 3 == TensorView::Rank,
+                       "Second View must be of tensors" );
+    }
+
+    /* Apply spline interpolation. */
+    template <class SplineDataType, class GridViewType>
+    KOKKOS_INLINE_FUNCTION void operator()( const SplineDataType& sd,
+                                            const int p,
+                                            const GridViewType& view ) const
+    {
+        scalar_value_type scalar_result;
+        Cajita::G2P::divergence( view, sd, scalar_result );
+        _x( p ) += scalar_result * 2.0;
+
+        tensor_value_type tensor_result[2][2];
+        Cajita::G2P::gradient( view, sd, tensor_result );
+
+        /* Thread-local manipulations may be computed here. */
+        for ( int i = 0; i < 2; ++i )
+            for ( int j = 0; j < 2; ++j )
+            {
+                _t( p, i, j ) = tensor_result[i][j];
+            }
+    }
+};
+
 //---------------------------------------------------------------------------//
 // Interpolation example.
[…]
