+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -210,6 +210,7 @@ void interpolationExample()
         num_point );
     auto tensor_point_host = Kokkos::create_mirror_view( tensor_point_field );
 
+    std::cout << "Invididual grid points before interpolation: "
     /***************************************************************************
      * P2G
      **************************************************************************/
@@ -217,9 +218,9 @@ void interpolationExample()
      * The Cajita::P2G namespace contains several methods for interpolating data
      * from particles to the grid. These interpolations are inherently scatter
      * operations for particle-based threading (a single particle maps to
-     * several grid nodes), which requires an underlying Kokkos::ScatterView for
-     * the data being interpolated. Of note, a thread-level interface provides
-     * methods to perform interpolations for a single particle datum to
+     * several grid points), which requires an underlying Kokkos::ScatterView
+     * for the data being interpolated. Of note, a thread-level interface
+     * provides methods to perform interpolations for a single particle datum to
      * neighboring mesh entitites.
      *
      * Cajita also provides a convenience interface for defining field-based P2G
@@ -256,44 +257,60 @@ void interpolationExample()
     Kokkos::deep_copy( vector_point_field, 3.5 );
     Kokkos::deep_copy( tensor_point_field, 3.5 );
 
-    // Interpolate a scalar point value to the grid.
+    // Reset the grid to zero.
     Cajita::ArrayOp::assign( *scalar_grid_field, 0.0, Cajita::Ghost() );
+    // Interpolate a scalar point value to the grid.
     auto scalar_p2g = Cajita::createScalarValueP2G( scalar_point_field, -0.5 );
     Cajita::p2g( scalar_p2g, points, num_point, Cajita::Spline<1>(),
                  *scalar_halo, *scalar_grid_field );
+    auto scalar_view = scalar_grid_field->view();
+    // Print out a random grid point.
+    std::cout << scalar_view( 5, 5, 5 );
 
-    // Interpolate a vector point value to the grid.
+    // Reset the grid to zero.
     Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
+    // Interpolate a vector point value to the grid.
     auto vector_p2g = Cajita::createVectorValueP2G( vector_point_field, -0.5 );
     Cajita::p2g( vector_p2g, points, num_point, Cajita::Spline<1>(),
                  *vector_halo, *vector_grid_field );
+    auto vector_view = vector_grid_field->view();
+    // Print out a random grid point.
+    std::cout << vector_grid_field->view()( 5, 5, 5 );
 
-    // Interpolate a scalar point gradient value to the grid.
+    // Reset the grid to zero.
     Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
+    // Interpolate a scalar point gradient value to the grid.
     auto scalar_grad_p2g =
         Cajita::createScalarGradientP2G( scalar_point_field, -0.5 );
     Cajita::p2g( scalar_grad_p2g, points, num_point, Cajita::Spline<1>(),
                  *vector_halo, *vector_grid_field );
+    // Print out a random grid point.
+    std::cout << vector_view( 5, 5, 5 );
 
-    // Interpolate a tensor point divergence value to the grid.
+    // Reset the grid to zero.
     Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
+    // Interpolate a tensor point divergence value to the grid.
     auto tensor_div_p2g =
         Cajita::createTensorDivergenceP2G( tensor_point_field, -0.5 );
     Cajita::p2g( tensor_div_p2g, points, num_point, Cajita::Spline<1>(),
                  *vector_halo, *vector_grid_field );
+    // Print out a random grid point.
[…]
