### 2022-04-04 #4115 @diehlpk
> Checking the BibTeX entries failed with the following error:
>
> ```
> Failed to parse BibTeX on value "doi" (NAME) [#<BibTeX::Bibliography data=[17]>, "@", #<BibTeX::Entry >, {:title=>["N ways to simulate short-range particle systems: Automated algorithm selection with the node-level library AutoPas"], :author=>["Gratl, Fabio Alexander and Seckler, Steffen and Bungartz, Hans-Joachim and Neumann, Philipp"], :journal=>["Computer Physics Communications"], :pages=>["108262"], :year=>["2021"], :publisher=>["Elsevier"]}]
> ```

@streeve Could you please check this new error?
