+++ paper/paper.bib
@@ -198,7 +198,7 @@
   journal={Computer Physics Communications},
   pages={108262},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi="10.1016/j.cpc.2021.108262"
 }
 
@@ -209,6 +209,6 @@
   volume={241},
   pages={155--177},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi="10.1016/j.cpc.2019.03.007"
 }
