### 2022-04-04 7daaaba
Added std::out for each set of interpolations.
