+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -210,7 +210,6 @@ void interpolationExample()
         num_point );
     auto tensor_point_host = Kokkos::create_mirror_view( tensor_point_field );
 
-    std::cout << "Invididual grid points before interpolation: "
     /***************************************************************************
      * P2G
      **************************************************************************/
@@ -252,6 +251,9 @@ void interpolationExample()
      * perform interpolations of particle data of various ranks.
      * *************************************************************************/
 
+    std::cout << "P2G interpolations" << std::endl;
+    std::cout << "-----------------------------------------------" << std::endl;
+
     // Initialize the particle data fields.
     Kokkos::deep_copy( scalar_point_field, 3.5 );
     Kokkos::deep_copy( vector_point_field, 3.5 );
@@ -259,43 +261,72 @@ void interpolationExample()
 
     // Reset the grid to zero.
     Cajita::ArrayOp::assign( *scalar_grid_field, 0.0, Cajita::Ghost() );
+    auto scalar_view = scalar_grid_field->view();
+    std::cout << "Invididual grid point scalar value at (5, 5, 5):\n\tbefore "
+                 "p2g::value interpolation: "
+              << scalar_view( 5, 5, 5 );
+
     // Interpolate a scalar point value to the grid.
     auto scalar_p2g = Cajita::createScalarValueP2G( scalar_point_field, -0.5 );
     Cajita::p2g( scalar_p2g, points, num_point, Cajita::Spline<1>(),
                  *scalar_halo, *scalar_grid_field );
-    auto scalar_view = scalar_grid_field->view();
+
     // Print out a random grid point.
-    std::cout << scalar_view( 5, 5, 5 );
+    std::cout << "\n\tafter p2g::value interpolation: "
+              << scalar_view( 5, 5, 5 ) << "\n\n";
 
     // Reset the grid to zero.
-    Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
-    // Interpolate a vector point value to the grid.
-    auto vector_p2g = Cajita::createVectorValueP2G( vector_point_field, -0.5 );
-    Cajita::p2g( vector_p2g, points, num_point, Cajita::Spline<1>(),
-                 *vector_halo, *vector_grid_field );
     auto vector_view = vector_grid_field->view();
-    // Print out a random grid point.
-    std::cout << vector_grid_field->view()( 5, 5, 5 );
+    std::cout << "Invididual grid point vector value at (5, 5, 5):\n\tbefore "
+                 "p2g::gradient interpolation: <";
+    std::cout << vector_view( 5, 5, 0 ) << ", " << vector_view( 5, 5, 1 )
+              << ">";
 
-    // Reset the grid to zero.
-    Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
     // Interpolate a scalar point gradient value to the grid.
     auto scalar_grad_p2g =
         Cajita::createScalarGradientP2G( scalar_point_field, -0.5 );
     Cajita::p2g( scalar_grad_p2g, points, num_point, Cajita::Spline<1>(),
                  *vector_halo, *vector_grid_field );
+
     // Print out a random grid point.
-    std::cout << vector_view( 5, 5, 5 );
+    std::cout << "\n\tafter p2g::gradient interpolation: <";
+    std::cout << vector_view( 5, 5, 0 ) << ", " << vector_view( 5, 5, 1 ) << ">"
+              << "\n\n";
 
     // Reset the grid to zero.
     Cajita::ArrayOp::assign( *vector_grid_field, 0.0, Cajita::Ghost() );
+    std::cout << "Invididual grid point vector value at (5, 5, 5):\n\tbefore "
+                 "p2g::divergence interpolation: <";
[…]
