+++ paper/paper.bib
@@ -199,7 +199,7 @@
   pages={108262},
   year={2021},
   publisher={Elsevier},
-  doi="10.1016/j.cpc.2021.108262"
+  doi={10.1016/j.cpc.2021.108262}
 }
 
 @article{openfpm,
@@ -210,5 +210,5 @@
   pages={155--177},
   year={2019},
   publisher={Elsevier},
-  doi="10.1016/j.cpc.2019.03.007"
+  doi={10.1016/j.cpc.2019.03.007}
 }
