### 2022-04-05 863f623
Update comments and remove deep copies
