+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -126,6 +126,8 @@ void interpolationExample()
       First, we need some setup to demonstrate the use of Cajita interpolation.
       This includes the creation of a simple uniform mesh and various fields on
       particles and the mesh.
+
+      This example is all in 2D.
     */
     using ExecutionSpace = Kokkos::DefaultHostExecutionSpace;
 
@@ -178,8 +180,6 @@ void interpolationExample()
     // detail in the Halo tutorial example.
     auto scalar_halo =
         Cajita::createHalo( *scalar_grid_field, Cajita::NodeHaloPattern<2>() );
-    auto scalar_grid_host =
-        Kokkos::create_mirror_view( scalar_grid_field->view() );
 
     // Create a vector field on the grid.
     auto vector_layout =
@@ -188,27 +188,22 @@ void interpolationExample()
         "vector_grid_field", vector_layout );
     auto vector_halo =
         Cajita::createHalo( *vector_grid_field, Cajita::NodeHaloPattern<2>() );
-    auto vector_grid_host =
-        Kokkos::create_mirror_view( vector_grid_field->view() );
 
     // Simple Kokkos::Views may be used to represent particle data.
     // Create a scalar point field.
     Kokkos::View<double*, ExecutionSpace> scalar_point_field(
         Kokkos::ViewAllocateWithoutInitializing( "scalar_point_field" ),
         num_point );
-    auto scalar_point_host = Kokkos::create_mirror_view( scalar_point_field );
 
     // Create a vector point field.
     Kokkos::View<double* [2], ExecutionSpace> vector_point_field(
         Kokkos::ViewAllocateWithoutInitializing( "vector_point_field" ),
         num_point );
-    auto vector_point_host = Kokkos::create_mirror_view( vector_point_field );
 
     // Create a tensor point field.
     Kokkos::View<double* [2][2], ExecutionSpace> tensor_point_field(
         Kokkos::ViewAllocateWithoutInitializing( "tensor_point_field" ),
         num_point );
-    auto tensor_point_host = Kokkos::create_mirror_view( tensor_point_field );
 
     /***************************************************************************
      * P2G
@@ -244,13 +239,6 @@ void interpolationExample()
      *  grid_field : A Cajita::Array on which to perform the scatter
      */
 
-    /***************************************************************************
-     * Cajita provides a basic set of P2G  and G2P functors and corresponding
-     * creation routines. The following examples demonstrate the use of these
-     * built-in functors in combination with Cajita::p2g() and Cajita::g2p() to
-     * perform interpolations of particle data of various ranks.
-     * *************************************************************************/
-
     std::cout << "P2G interpolations" << std::endl;
     std::cout << "-----------------------------------------------" << std::endl;
 
@@ -262,22 +250,31 @@ void interpolationExample()
     // Reset the grid to zero.
     Cajita::ArrayOp::assign( *scalar_grid_field, 0.0, Cajita::Ghost() );
     auto scalar_view = scalar_grid_field->view();
-    std::cout << "Invididual grid point scalar value at (5, 5, 5):\n\tbefore "
+
+    // Print out a random grid point before value interpolation.
+    std::cout << "Invididual grid point at (5, 5, 5):\n\tbefore "
                  "p2g::value interpolation: "
               << scalar_view( 5, 5, 5 );
 
-    // Interpolate a scalar point value to the grid.
+    /*
+      Cajita provides a basic set of P2G and G2P functors and corresponding
+      creation routines. The next four interpolations below demonstrate the use
+      of these built-in functors in combination with the Cajita::p2g(), and
+      later Cajita::g2p().
+    */
[…]
