+++ example/cajita_tutorial/15_interpolation/interpolation_example.cpp
@@ -50,19 +50,19 @@ struct P2GExampleFunctor
                                             const GridViewType& view ) const
     {
         /* Access the point data */
-        scalar_value_type scalar_point_data = _x( p );
-        vector_value_type vector_point_data[2];
+        scalar_value_type scalar_particle_data = _x( p );
+        vector_value_type vector_particle_data[2];
 
         /* Thread-local manipulations may be computed here. */
         for ( int d = 0; d < 2; ++d )
         {
-            vector_point_data[d] = 3.0 * _y( p, d );
+            vector_particle_data[d] = 3.0 * _y( p, d );
         }
-        scalar_point_data *= 2.0;
+        scalar_particle_data *= 2.0;
 
         /* Finally, call the thread-level value spline interpolations. */
-        Cajita::P2G::value( scalar_point_data, sd, view );
-        Cajita::P2G::divergence( vector_point_data, sd, view );
+        Cajita::P2G::value( scalar_particle_data, sd, view );
+        Cajita::P2G::divergence( vector_particle_data, sd, view );
     }
 };
 
@@ -152,11 +152,11 @@ void interpolationExample()
     // Create a point in the center of every cell.
     auto cell_space = local_grid->indexSpace( Cajita::Own(), Cajita::Cell(),
                                               Cajita::Local() );
-    int num_point = cell_space.size();
-    Kokkos::View<double* [2], ExecutionSpace> points(
-        Kokkos::ViewAllocateWithoutInitializing( "points" ), num_point );
+    int num_particles = cell_space.size();
+    Kokkos::View<double* [2], ExecutionSpace> particle_positions(
+        Kokkos::ViewAllocateWithoutInitializing( "positions" ), num_particles );
     Kokkos::parallel_for(
-        "fill_points", createExecutionPolicy( cell_space, ExecutionSpace() ),
+        "fill_positions", createExecutionPolicy( cell_space, ExecutionSpace() ),
         KOKKOS_LAMBDA( const int i, const int j ) {
             int pi = i - halo_width;
             int pj = j - halo_width;
@@ -164,8 +164,8 @@ void interpolationExample()
             int idx[2] = { i, j };
             double x[2];
             local_mesh.coordinates( Cajita::Cell(), idx, x );
-            points( pid, Cajita::Dim::I ) = x[Cajita::Dim::I];
-            points( pid, Cajita::Dim::J ) = x[Cajita::Dim::J];
+            particle_positions( pid, Cajita::Dim::I ) = x[Cajita::Dim::I];
+            particle_positions( pid, Cajita::Dim::J ) = x[Cajita::Dim::J];
         } );
 
     // Next, we use Cajita functionality to create grid data fields.
@@ -191,19 +191,19 @@ void interpolationExample()
 
     // Simple Kokkos::Views may be used to represent particle data.
     // Create a scalar point field.
-    Kokkos::View<double*, ExecutionSpace> scalar_point_field(
-        Kokkos::ViewAllocateWithoutInitializing( "scalar_point_field" ),
-        num_point );
+    Kokkos::View<double*, ExecutionSpace> scalar_particle_field(
+        Kokkos::ViewAllocateWithoutInitializing( "particle_scalar" ),
+        num_particles );
 
     // Create a vector point field.
-    Kokkos::View<double* [2], ExecutionSpace> vector_point_field(
-        Kokkos::ViewAllocateWithoutInitializing( "vector_point_field" ),
-        num_point );
+    Kokkos::View<double* [2], ExecutionSpace> vector_particle_field(
+        Kokkos::ViewAllocateWithoutInitializing( "particle_vector" ),
[…]
