### 2022-04-05 #4115 @diehlpk
@streeve Looks good.

Please do the following to finish the paper

- [ ] Generate a new version, including all changes during the JOSS review, and post the new version number here.
- [ ] Upload this version to Zenodo or FigShare and post the DOI here. Note that the title and authors of the paper have to match with the authors and title of the archive. Please post the DOI here.
