+++ doxygen/search/functions_d.js
@@ -1,20 +1,20 @@
 var searchData=
 [
-  ['neighbor_5fparallel_5ffor_1228',['neighbor_parallel_for',['../namespaceCabana.html#acb924882c14a90b695a56065f4106c22',1,'Cabana::neighbor_parallel_for(const Kokkos::RangePolicy&lt; ExecParameters... &gt; &amp;exec_policy, const FunctorType &amp;functor, const NeighborListType &amp;list, const FirstNeighborsTag, const SerialOpTag, const std::string &amp;str=&quot;&quot;)'],['../namespaceCabana.html#a10eeaa254bbc26461eadae4d9a0a1598',1,'Cabana::neighbor_parallel_for(const Kokkos::RangePolicy&lt; ExecParameters... &gt; &amp;exec_policy, const FunctorType &amp;functor, const NeighborListType &amp;list, const SecondNeighborsTag, const SerialOpTag, const std::string &amp;str=&quot;&quot;)'],['../namespaceCabana.html#ac2cb6955452842c7903e86a845329e7c',1,'Cabana::neighbor_parallel_for(const Kokkos::RangePolicy&lt; ExecParameters... &gt; &amp;exec_policy, const FunctorType &amp;functor, const NeighborListType &amp;list, const FirstNeighborsTag, const TeamOpTag, const std::string &amp;str=&quot;&quot;)'],['../namespaceCabana.html#acca1d55078d6cfea288c6cec130ef204',1,'Cabana::neighbor_parallel_for(const Kokkos::RangePolicy&lt; ExecParameters... &gt; &amp;exec_policy, const FunctorType &amp;functor, const NeighborListType &amp;list, const SecondNeighborsTag, const TeamOpTag, const std::string &amp;str=&quot;&quot;)'],['../namespaceCabana.html#a3dae9114f1f2ede7f1925a0c73c7639b',1,'Cabana::neighbor_parallel_for(const Kokkos::RangePolicy&lt; ExecParameters... &gt; &amp;exec_policy, const FunctorType &amp;functor, const NeighborListType &amp;list, const SecondNeighborsTag, const TeamVectorOpTag, const std::string &amp;str=&quot;&quot;)']]],
[…]
