+++ doxygen/search/functions_a.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['key2ijk_1201',['key2ijk',['../classCajita_1_1SparseMap.html#a61e4bcbc0b530abe6ae3682cfb9f0f2b',1,'Cajita::SparseMap::key2ijk()'],['../classCajita_1_1BlockMap.html#a18a36518c0c426cd1b13b4ab6236c5ec',1,'Cajita::BlockMap::key2ijk()']]],
-  ['key_5fat_1202',['key_at',['../classCajita_1_1SparseMap.html#ac21f0acbf87da9f7306a803710ffa913',1,'Cajita::SparseMap::key_at()'],['../classCajita_1_1BlockMap.html#ab373ab6ebe66918e8348769ec610d529',1,'Cajita::BlockMap::key_at()']]],
-  ['keyminmax_1203',['keyMinMax',['../Cabana__Sort_8hpp.html#a2477c75803a67977e5d1f9f59ab08396',1,'Cabana::Impl']]],
-  ['kokkosbinsort_1204',['kokkosBinSort',['../Cabana__Sort_8hpp.html#a67c1dd70c87d16e74ade337b737d810d',1,'Cabana::Impl']]],
-  ['kokkosbinsort1d_1205',['kokkosBinSort1d',['../Cabana__Sort_8hpp.html#a8ea3d9a71899ddf9b07f8831b6457c61',1,'Cabana::Impl']]]
+  ['key2ijk_1204',['key2ijk',['../classCajita_1_1SparseMap.html#a61e4bcbc0b530abe6ae3682cfb9f0f2b',1,'Cajita::SparseMap::key2ijk()'],['../classCajita_1_1BlockMap.html#a18a36518c0c426cd1b13b4ab6236c5ec',1,'Cajita::BlockMap::key2ijk()']]],
+  ['key_5fat_1205',['key_at',['../classCajita_1_1SparseMap.html#ac21f0acbf87da9f7306a803710ffa913',1,'Cajita::SparseMap::key_at()'],['../classCajita_1_1BlockMap.html#ab373ab6ebe66918e8348769ec610d529',1,'Cajita::BlockMap::key_at()']]],
+  ['keyminmax_1206',['keyMinMax',['../Cabana__Sort_8hpp.html#a2477c75803a67977e5d1f9f59ab08396',1,'Cabana::Impl']]],
+  ['kokkosbinsort_1207',['kokkosBinSort',['../Cabana__Sort_8hpp.html#a67c1dd70c87d16e74ade337b737d810d',1,'Cabana::Impl']]],
+  ['kokkosbinsort1d_1208',['kokkosBinSort1d',['../Cabana__Sort_8hpp.html#a8ea3d9a71899ddf9b07f8831b6457c61',1,'Cabana::Impl']]]
 ];
