+++ doxygen/search/variables_b.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['periodic_1414',['periodic',['../structCajita_1_1IndexConversion_1_1L2G.html#a467de8ddb7856a73d2a6d2b8626b7bfe',1,'Cajita::IndexConversion::L2G']]]
+  ['periodic_1419',['periodic',['../structCajita_1_1IndexConversion_1_1L2G.html#a467de8ddb7856a73d2a6d2b8626b7bfe',1,'Cajita::IndexConversion::L2G']]]
 ];
