+++ doxygen/search/variables_10.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['w_1432',['w',['../structCajita_1_1SplineDataMember_3_01Scalar_00_01Order_00_01NumSpaceDim_00_01SplineWeightValues_01_4.html#a28cd8136a08f7b90c8295bb43a6a86fd',1,'Cajita::SplineDataMember&lt; Scalar, Order, NumSpaceDim, SplineWeightValues &gt;']]],
-  ['workload_5fprefix_5fsum_1433',['workload_prefix_sum',['../structCajita_1_1SparseDimPartitioner_1_1SubWorkloadFunctor.html#a0ead344673c61cb0926bb57be3d9bbc3',1,'Cajita::SparseDimPartitioner::SubWorkloadFunctor']]]
+  ['w_1437',['w',['../structCajita_1_1SplineDataMember_3_01Scalar_00_01Order_00_01NumSpaceDim_00_01SplineWeightValues_01_4.html#a28cd8136a08f7b90c8295bb43a6a86fd',1,'Cajita::SplineDataMember&lt; Scalar, Order, NumSpaceDim, SplineWeightValues &gt;']]],
+  ['workload_5fprefix_5fsum_1438',['workload_prefix_sum',['../structCajita_1_1SparseDimPartitioner_1_1SubWorkloadFunctor.html#a6e404714b157f7db41e50563c553c566',1,'Cajita::SparseDimPartitioner::SubWorkloadFunctor']]]
 ];

[…]
