+++ cajita/src/Cajita_SparseDimPartitioner.hpp
@@ -73,8 +73,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
     /*!
       \brief Constructor - automatically compute ranks_per_dim from MPI
       communicator
-      \param comm MPI communicator to decide the rank nums in each
-      dimention
+      \param comm MPI communicator to decide the rank nums in each dimension
       \param max_workload_coeff threshold factor for re-partition
       \param workload_num total workload(particle/tile) number, used to compute
       workload_threshold
@@ -101,17 +100,20 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
     /*!
       \brief Constructor - user-defined ranks_per_dim
       communicator
+      \param comm MPI communicator to decide the rank nums in each dimension
       \param max_workload_coeff threshold factor for re-partition
       \param workload_num total workload(particle/tile) number, used to compute
       workload_threshold
       \param num_step_rebalance the simulation step number after which one
       should check if repartition is needed
-      \param ranks_per_dim 3D array, user-defined MPI ranks in per dimension
+      \param ranks_per_dim 3D array, user-defined MPI rank constrains in per
+      dimension
       \param global_cells_per_dim 3D array, global cells in each dimension
       \param max_optimize_iteration max iteration number to run the optimization
     */
     SparseDimPartitioner(
-        float max_workload_coeff, int workload_num, int num_step_rebalance,
+        MPI_Comm comm, float max_workload_coeff, int workload_num,
+        int num_step_rebalance,
         const std::array<int, num_space_dim>& ranks_per_dim,
         const std::array<int, num_space_dim>& global_cells_per_dim,
         int max_optimize_iteration = 10 )
@@ -123,6 +125,11 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
         allocate( global_cells_per_dim );
         std::copy( ranks_per_dim.begin(), ranks_per_dim.end(),
                    _ranks_per_dim.data() );
+
+        // init MPI topology
+        int comm_size;
+        MPI_Comm_size( comm, &comm_size );
+        MPI_Dims_create( comm_size, num_space_dim, _ranks_per_dim.data() );
     }
 
     /*!
@@ -367,6 +374,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
                          cell_bits_per_tile_dim;
                 Kokkos::atomic_increment( &workload( ti + 1, tj + 1, tz + 1 ) );
             } );
+        Kokkos::fence();
     }
 
     /*!
@@ -393,6 +401,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
                         &workload( ti + 1, tj + 1, tk + 1 ) );
                 }
             } );
+        Kokkos::fence();
     }
 
     /*!
@@ -411,6 +420,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
         // workload matrix, save the results in _workload_prefix_sum
         MPI_Allreduce( workload.data(), prefix_sum.data(), total_size, MPI_INT,
                        MPI_SUM, comm );
+        MPI_Barrier( comm );
 
         // compute the prefix sum (in three dimensions)
         // prefix sum in the dimension 0
@@ -432,6 +442,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
                             prefix_sum( i, j, k ) = update;
                         }
                     } );
+        Kokkos::fence();
 
         // prefix sum in the dimension 1
         for ( int i = 0;
@@ -452,6 +463,7 @@ class SparseDimPartitioner : public BlockPartitioner<NumSpaceDim>
                             prefix_sum( i, j, k ) = update;
[…]
