### 2022-04-06 0080e6c
make cajita reference solver simple to be tested on owned space
