### 2022-04-06 4a5e776
delete unused variable
