### 2022-04-06 a7d8596
Match details between reference and hypre solver examples
