### 2022-04-06 #4115 @streeve
Things are nearly all merged to do the release. I'm not sure I understand the second bullet - I would normally just title "Cabana Version 0.5". Does that code release need to be done manually rather than through Github in order to match the title and author list?
