### 2022-04-06 #4115 @diehlpk
No, the title of this release must be "Cabana: A Performance Portable Library for Particle-Based Simulations" like on the paper. And the same list of authors. You can do just a normal release as usual with the title "Cabana Version 0.5". In Addition, you upload the content of this release to Zenodo or FigShare with the paper's title.
