### 2022-04-08 beea11d
Workaround for FindGTest module with cmake version < 3.20
