### 2022-04-08 78266c9
Fixup gtest find module workaround (cannot alias imported target)
